### 2025-07-03 #8536 @editorialbot
[PRE REVIEW]: chatAI4R: Interactive Artificial Intelligence toolkit for Data Science in R
**Submitting author:** @kumeS (<a href="http://orcid.org/0000-0001-7481-2843">Satoshi Kume</a>)
**Repository:** https://github.com/kumeS/chatAI4R
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
