### 2025-07-16 kumeS/chatAI4BI#9 @kumeS
JOSS review 対策
下記に対処してください。/paper/paper.mdにおけるライセンスの件です。

License info:

https://github.com/openjournals/joss-reviews/issues/8536#issuecomment-3032454091
