### 2025-07-16 kumeS/chatAI4R#9 @kumeS
@claude なぜfailedしますか？
