### 2025-07-30 f445975
Add multi-provider API support to addCommentCode function
- Add support for both OpenAI and Gemini API keys
- Implement auto-detection of available API providers
- Add provider parameter with "auto", "openai", "gemini" options
- Route API calls to appropriate provider (chat4R_history or gemini4R)
- Update function documentation with new parameter and examples
- Resolves issue with hardcoded OPENAI_API_KEY requirement

Fixes #11

Co-authored-by: skume <kumeS@users.noreply.github.com>
