+++ R/addCommentCode.R
@@ -3,9 +3,10 @@
 #' @title Add Comments to R Code
 #' @description This function adds comments to R code without modifying the input R code.
 #'    It can either take the selected code from RStudio or read from the clipboard.
-#' @param Model A character string specifying the GPT model to be used. Default is "gpt-4o-mini".
+#' @param Model A character string specifying the model to be used. Default is "gpt-4o-mini" for OpenAI or "gemini-2.0-flash" for Gemini.
 #' @param language A character string specifying the language for the comments. Default is "English".
 #' @param SelectedCode A logical value indicating whether to use the selected code in RStudio. Default is TRUE.
+#' @param provider A character string specifying the API provider. Options: "auto" (default), "openai", "gemini". When "auto", automatically detects available API keys.
 #' @importFrom assertthat assert_that is.string noNA
 #' @importFrom clipr read_clip write_clip
 #' @importFrom rstudioapi isAvailable getActiveDocumentContext insertText
@@ -15,16 +16,24 @@
 #' @author Satoshi Kume
 #' @examples
 #' \dontrun{
-#' # Option 1
+#' # Option 1: Auto-detect available API (OpenAI or Gemini)
 #' # Select some text in RStudio and then run the rstudio addins
-#' # Option 2
-#' # Copy the text into your clipboard then execute
 #' addCommentCode(Model = "gpt-4o-mini", language = "English", SelectedCode = TRUE)
+#' 
+#' # Option 2: Explicitly use OpenAI
+#' addCommentCode(Model = "gpt-4o-mini", provider = "openai", SelectedCode = TRUE)
+#' 
+#' # Option 3: Use Gemini API
+#' addCommentCode(Model = "gemini-2.0-flash", provider = "gemini", SelectedCode = TRUE)
+#' 
+#' # Option 4: Copy text to clipboard and execute
+#' addCommentCode(Model = "gpt-4o-mini", language = "English", SelectedCode = FALSE)
 #' }
 
 addCommentCode <- function(Model = "gpt-4o-mini",
                            language = "English",
-                           SelectedCode = TRUE) {
+                           SelectedCode = TRUE,
+                           provider = "auto") {
 
   # Get input either from RStudio or clipboard
   if(SelectedCode){
@@ -34,12 +43,36 @@ addCommentCode <- function(Model = "gpt-4o-mini",
     input <- paste0(clipr::read_clip(), collapse = " \n")
   }
 
+  # Detect available API providers
+  openai_available <- Sys.getenv("OPENAI_API_KEY") != ""
+  gemini_available <- Sys.getenv("GoogleGemini_API_KEY") != ""
+  
+  # Auto-detect provider if not specified
+  if (provider == "auto") {
+    if (openai_available) {
+      provider <- "openai"
+    } else if (gemini_available) {
+      provider <- "gemini"
+    } else {
+      stop("No API keys found. Set OPENAI_API_KEY or GoogleGemini_API_KEY environment variable.")
+    }
+  }
+  
+  # Validate selected provider has API key
+  if (provider == "openai" && !openai_available) {
+    stop("OpenAI API key not found. Set OPENAI_API_KEY environment variable.")
+  }
+  if (provider == "gemini" && !gemini_available) {
+    stop("Gemini API key not found. Set GoogleGemini_API_KEY environment variable.")
+  }
+  
   # Assertions for input validation
   assertthat::assert_that(
     assertthat::is.string(input),
     assertthat::noNA(input),
     assertthat::is.string(Model),
-    Sys.getenv("OPENAI_API_KEY") != ""
+    assertthat::is.string(provider),
+    provider %in% c("openai", "gemini")
   )
 
   # Initialize temperature for text generation
[…]
