### 2025-07-30 2ac8955
Add multi-provider API support to addCommentCode function
- Add support for OpenAI, Gemini, and Vertex AI API providers
- Implement auto-detection of available API keys with fallback priority
- Add provider parameter with "auto", "openai", "gemini", "vertex" options
- Route to appropriate API functions based on provider selection
- Maintain backwards compatibility with existing OpenAI usage
- Add comprehensive documentation with usage examples
- Add proper error handling and validation for all providers

Resolves #11 - addCommentCode SUPPORTS ONLY Open AI API Key

🤖 Generated with [Claude Code
[…]
