+++ R/addCommentCode.R
@@ -3,28 +3,41 @@
 #' @title Add Comments to R Code
 #' @description This function adds comments to R code without modifying the input R code.
 #'    It can either take the selected code from RStudio or read from the clipboard.
-#' @param Model A character string specifying the GPT model to be used. Default is "gpt-4o-mini".
+#'    Supports multiple AI providers: OpenAI, Google Gemini, and Google Vertex AI.
+#' @param Model A character string specifying the AI model to be used. Default is "gpt-4o-mini".
+#'    Examples: "gpt-4o-mini", "gemini-2.0-flash", "gemini-1.5-pro" etc.
 #' @param language A character string specifying the language for the comments. Default is "English".
 #' @param SelectedCode A logical value indicating whether to use the selected code in RStudio. Default is TRUE.
+#' @param provider A character string specifying the AI provider. Options: "auto", "openai", "gemini", "vertex".
+#'    Default is "auto" which automatically detects available API keys (OpenAI -> Gemini -> Vertex AI).
 #' @importFrom assertthat assert_that is.string noNA
 #' @importFrom clipr read_clip write_clip
 #' @importFrom rstudioapi isAvailable getActiveDocumentContext insertText
+#' @importFrom httr POST add_headers content http_error status_code
+#' @importFrom jsonlite toJSON fromJSON
 #' @return A message indicating completion if `SelectedCode` is TRUE, otherwise the commented code is copied to the clipboard.
 #' @export addCommentCode
 #'
 #' @author Satoshi Kume
 #' @examples
 #' \dontrun{
-#' # Option 1
-#' # Select some text in RStudio and then run the rstudio addins
-#' # Option 2
-#' # Copy the text into your clipboard then execute
-#' addCommentCode(Model = "gpt-4o-mini", language = "English", SelectedCode = TRUE)
+#' # Auto-detect provider (OpenAI -> Gemini -> Vertex AI)
+#' addCommentCode(Model = "gpt-4o-mini", SelectedCode = TRUE)
+#' 
+#' # Explicitly use OpenAI
+#' addCommentCode(Model = "gpt-4o-mini", provider = "openai", SelectedCode = TRUE)
+#' 
+#' # Use Google Gemini
+#' addCommentCode(Model = "gemini-2.0-flash", provider = "gemini", SelectedCode = TRUE)
+#' 
+#' # Use Google Vertex AI
+#' addCommentCode(Model = "gemini-1.5-pro", provider = "vertex", SelectedCode = TRUE)
 #' }
 
 addCommentCode <- function(Model = "gpt-4o-mini",
                            language = "English",
-                           SelectedCode = TRUE) {
+                           SelectedCode = TRUE,
+                           provider = "auto") {
 
   # Get input either from RStudio or clipboard
   if(SelectedCode){
@@ -39,8 +52,43 @@ addCommentCode <- function(Model = "gpt-4o-mini",
     assertthat::is.string(input),
     assertthat::noNA(input),
     assertthat::is.string(Model),
-    Sys.getenv("OPENAI_API_KEY") != ""
+    assertthat::is.string(provider),
+    provider %in% c("auto", "openai", "gemini", "vertex")
   )
+
+  # Helper function to check API keys
+  check_api_keys <- function() {
+    list(
+      openai = Sys.getenv("OPENAI_API_KEY") != "",
+      gemini = Sys.getenv("GoogleGemini_API_KEY") != "",
+      vertex = Sys.getenv("GOOGLE_CLOUD_PROJECT") != "" && 
+               (Sys.getenv("GOOGLE_ACCESS_TOKEN") != "" || system("which gcloud", ignore.stderr = TRUE) == 0)
+    )
+  }
+
+  # Determine which provider to use
+  available_keys <- check_api_keys()
+  
[…]
