### 2025-07-31 8946542
v1.0.x: JOSS version
