### 2025-08-04 #8536 @kumeS
@crvernon

Dear Chris Vernon,

I believe the paper is ready for review.

I noticed your name mentioned in the first comment of the issue, and I was wondering if I should wait before proceeding with the review.

Best regards,
Satoshi Kume
