### 2025-08-19 48c78aa
feat: Add dynamic model fetching to multiLLMviaionet
- Implement dynamic model list retrieval from io.net API
- Add intelligent caching (1 hour) to improve performance
- Maintain backward compatibility with static fallback
- Add refresh_ionet_models() function for manual cache refresh
- Update list_ionet_models() with force_refresh option
- Handle API failures gracefully with fallback to static list
- Improve documentation with usage examples

Fixes #18

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-authored-by: skume <kumeS@users.noreply.github.com>
