+++ R/multiLLMviaionet.R
@@ -5,6 +5,10 @@
 #'   simultaneously using the io.net API. It supports up to 6 models running in parallel,
 #'   with streaming responses and comprehensive error handling.
 #'   
+#'   The function dynamically fetches the current list of available models from the io.net API
+#'   to ensure compatibility when model endpoints change. Results are cached for 1 hour to 
+#'   improve performance. If the API is unavailable, it falls back to a static model list.
+#'   
 #'   Supported model categories include:
 #'   - Meta Llama series (Llama-4-Maverick, Llama-3.3-70B, etc.)
 #'   - DeepSeek series (DeepSeek-R1, DeepSeek-R1-Distill variants)
@@ -28,7 +32,7 @@
 #' @param parallel Logical indicating whether to run models in parallel. Default is TRUE.
 #' @param verbose Logical indicating whether to show detailed progress. Default is TRUE.
 #'
-#' @importFrom httr POST add_headers content status_code
+#' @importFrom httr POST GET add_headers content status_code timeout
 #' @importFrom jsonlite toJSON fromJSON
 #' @importFrom parallel makeCluster stopCluster parLapply detectCores
 #' @importFrom future future multisession plan resolved value
@@ -47,12 +51,19 @@
 #'   # Set your io.net API key
 #'   Sys.setenv(IONET_API_KEY = "your_ionet_api_key_here")
 #'   
-#'   # Basic usage with default models
+#'   # Basic usage with default models (dynamically fetched)
 #'   result <- multiLLMviaionet(
 #'     prompt = "Explain quantum computing in simple terms",
 #'     models = c("meta-llama/Llama-3.3-70B-Instruct", 
 #'                "deepseek-ai/DeepSeek-R1",
 #'                "Qwen/Qwen3-235B-A22B-FP8")
+#'   )
+#'   
+#'   # Get current available models first, then use them
+#'   current_models <- refresh_ionet_models()
+#'   result <- multiLLMviaionet(
+#'     prompt = "What is machine learning?",
+#'     models = current_models[1:3]  # Use first 3 available models
 #'   )
 #'   
 #'   # Advanced usage with random selection (10 models)
@@ -119,8 +130,8 @@ multiLLMviaionet <- function(prompt,
     cat("** Prompt:", substr(prompt, 1, 100), if(nchar(prompt) > 100) "..." else "", "\n")
   }
   
-  # Get list of available models
-  available_models <- .get_ionet_available_models()
+  # Get list of available models (with dynamic API fetching)
+  available_models <- .get_ionet_available_models(api_key = api_key, verbose = verbose)
   
   # Handle random selection - override models parameter if random_selection is TRUE
   if (random_selection) {
@@ -186,8 +197,135 @@ multiLLMviaionet <- function(prompt,
   return(output)
 }
 
-# Helper function: Get available io.net models
-.get_ionet_available_models <- function() {
+# Helper function: Get available io.net models dynamically
+.get_ionet_available_models <- function(api_key = NULL, force_refresh = FALSE, verbose = FALSE) {
+  
+  # Use provided api_key or get from environment
+  if (is.null(api_key)) {
+    api_key <- Sys.getenv("IONET_API_KEY")
+  }
+  
+  # Check if we should use cached results
+  cache_key <- "ionet_models_cache"
+  cache_time_key <- "ionet_models_cache_time"
+  cache_duration <- 3600  # Cache for 1 hour
+  
+  # Get cached data if available and not forcing refresh
+  if (!force_refresh && exists(cache_key, envir = .GlobalEnv) && exists(cache_time_key, envir = .GlobalEnv)) {
+    cached_time <- get(cache_time_key, envir = .GlobalEnv)
+    if (as.numeric(difftime(Sys.time(), cached_time, units = "secs")) < cache_duration) {
[…]
