### 2025-10-27 kumeS/chatAI4R#24 @khoa-yelo
[JOSS Review] Installation
Could you take a look at this package installation error. Thanks

```bash
** R
** inst
** byte-compile and prepare package for lazy loading
** help
*** installing help indices
*** copying figures
** building package indices
** installing vignettes
** testing if installed package can be loaded from temporary location
** checking absolute paths in shared objects and dynamic libraries
** testing if installed package can be loaded from final location
** testing if installed package keeps a record of temporary installation path
* DONE (igraph)
* installing *source* package ‘deepRstudio’ ...
** package ‘deepRstudio’ successfully unpacked and MD5 sums checked
** using staged installation
** R
** inst
** byte-compile and prepare package for lazy loading
** help
*** installing help indices
** building package indices
** testing if installed package can be loaded from temporary location
** testing if installed package can be loaded from final location
** testing if installed package keeps a record of temporary installation path
* DONE (deepRstudio)
ERROR: dependency ‘qpdf’ is not available for package ‘pdftools’
* removing ‘/home/users/khoang99/R/x86_64-pc-linux-gnu-library/4.3/pdftools’
Warning in install.packages :
  installation of package ‘pdftools’ had non-zero exit status
ERROR: dependency ‘pdftools’ is not available for package ‘chatAI4R’
* removing ‘/home/users/khoang99/R/x86_64-pc-linux-gnu-library/4.3/chatAI4R’
Warning in install.packages :
  installation of package ‘chatAI4R’ had non-zero exit status
```
