### 2025-10-27 #9283 @khoa-yelo
@kumeS Could you take a look at this installation issue, thanks!
https://github.com/kumeS/chatAI4R/issues/24
