### 2025-10-27 07a4796
docs: add system dependencies installation instructions
Add qpdf system dependency installation instructions to prevent
installation errors like the one reported in issue #24.

The pdftools R package requires qpdf system library to be installed
before the chatAI4R package can be successfully installed.

Co-authored-by: skume <kumeS@users.noreply.github.com>
