### 2025-10-27 kumeS/chatAI4R#24 @kumeS
Hi @khoa-yelo ,
Please choose Option 1: Install dependencies manually, excluding pdftools, on your linux system.
