### 2025-10-27 kumeS/chatAI4R#24 @kumeS
@khoa-yelo OK, thank you very much for your confirmation. I will check it again. Please give me time to re-built the package.
