### 2025-10-27 34214d0
fix: make pdftools an optional dependency for Linux compatibility
Move pdftools from Imports to Suggests to resolve installation issues
on Linux systems where libqpdf-dev may not be available. This allows
users to install chatAI4R without pdftools and only install it when
needed for PDF processing functionality.

Changes:
- DESCRIPTION: Move pdftools from Imports to Suggests
- R/chatAI4pdf.R: Add runtime check for pdftools with helpful error message
- NAMESPACE: Remove pdftools import (auto-regenerated)
- man/chatAI4pdf.Rd: Update documentation with Linux dependency notes

Fixes Linux installa
[…]
