+++ NAMESPACE
@@ -36,12 +36,11 @@ export(gemini4R)
 export(geminiGrounding4R)
 export(interpretResult)
 export(list_ionet_models)
-export(multiLLM_random10)
-export(multiLLM_random5)
 export(multiLLMviaionet)
 export(ngsub)
 export(proofreadEnglishText)
 export(proofreadText)
+export(refresh_ionet_models)
 export(removeQuotations)
 export(replicatellmAPI4R)
 export(revisedText)
@@ -83,9 +82,11 @@ importFrom(httr,POST)
 importFrom(httr,add_headers)
 importFrom(httr,content)
 importFrom(httr,status_code)
+importFrom(httr,timeout)
 importFrom(httr,write_stream)
 importFrom(igraph,add_edges)
 importFrom(igraph,add_vertices)
+importFrom(igraph,graph)
 importFrom(igraph,layout_nicely)
 importFrom(igraph,layout_with_fr)
 importFrom(jsonlite,fromJSON)
@@ -94,7 +95,6 @@ importFrom(parallel,detectCores)
 importFrom(parallel,makeCluster)
 importFrom(parallel,parLapply)
 importFrom(parallel,stopCluster)
-importFrom(pdftools,pdf_text)
 importFrom(rstudioapi,getActiveDocumentContext)
 importFrom(rstudioapi,insertText)
 importFrom(rstudioapi,isAvailable)
