### 2025-10-27 601b745
v1.0.x: JOSS version
