+++ man/multiLLM_random5.Rd
@@ -1,27 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/multiLLMviaionet.R
-\name{multiLLM_random5}
-\alias{multiLLM_random5}
-\title{multiLLM_random5: Quick execution of 5 randomly selected models}
-\usage{
-multiLLM_random5(prompt, ...)
-}
-\arguments{
-\item{prompt}{A string containing the input prompt.}
-
-\item{...}{Additional arguments passed to multiLLM_random10()}
-}
-\value{
-A multiLLM_result object with results from 5 randomly selected models.
-}
-\description{
-Convenience function for faster testing with 5 diverse models.
-}
-\details{
-Quick 5-Model Random Selection
-}
-\examples{
-\dontrun{
-  result <- multiLLM_random5("What is artificial intelligence?")
-}
-}
