### 2025-10-27 24e2fca
docs: add comprehensive Linux installation guide for pdftools dependency
Addresses Linux installation issues where users encounter "dependency 'pdftools' is not available" error.

Changes:
- Enhanced README.md with prominent Linux-specific installation section
- Added step-by-step instructions for Ubuntu/Debian, CentOS/RHEL, Rocky Linux/Fedora
- Documented alternative installation method without PDF functionality
- Created dedicated LINUX_INSTALL.md with comprehensive troubleshooting guide
- Added verification steps and system requirements documentation
- Included instructions for both full in
[…]
