+++ LINUX_INSTALL.md
@@ -1,0 +1,284 @@
+# Linux Installation Guide for chatAI4R
+
+## 🐧 Quick Start for Linux Users
+
+This guide provides step-by-step instructions for installing **chatAI4R** on Linux systems.
+
+---
+
+## Problem: "dependency 'pdftools' is not available"
+
+If you see this error:
+```
+ERROR: dependency 'pdftools' is not available for package 'chatAI4R'
+```
+
+This happens because the `pdftools` R package requires system libraries that aren't installed by default on Linux.
+
+---
+
+## ✅ Solution 1: Install with Full PDF Support (Recommended)
+
+Follow these steps to install chatAI4R with complete PDF processing functionality:
+
+### Step 1: Install System Dependencies
+
+Choose your Linux distribution:
+
+#### **Ubuntu / Debian / Linux Mint**
+```bash
+sudo apt-get update
+sudo apt-get install -y libqpdf-dev
+```
+
+#### **CentOS / RHEL (versions 7-8)**
+```bash
+sudo yum install -y qpdf-devel
+```
+
+#### **Rocky Linux / AlmaLinux / Fedora (newer versions)**
+```bash
+sudo dnf install -y qpdf-devel
+```
+
+#### **Arch Linux / Manjaro**
+```bash
+sudo pacman -S qpdf
+```
+
+#### **openSUSE**
+```bash
+sudo zypper install qpdf-devel
+```
+
+### Step 2: Verify System Library Installation
+
+```bash
+# Check if qpdf library is installed
+ldconfig -p | grep qpdf
+```
+
+You should see output like:
+```
+libqpdf.so.29 (libc6,x86-64) => /usr/lib/x86_64-linux-gnu/libqpdf.so.29
+```
+
+### Step 3: Install pdftools in R
+
+Open R and run:
+```r
+# Install pdftools first
+install.packages("pdftools")
+
+# Verify installation
+library(pdftools)
+```
+
+If successful, you'll see no errors.
+
+### Step 4: Install chatAI4R
+
+```r
+# Now install chatAI4R with all dependencies
+install.packages("chatAI4R")
+
+# Load the package
+library(chatAI4R)
+```
+
+---
+
+## ⚡ Solution 2: Install Without PDF Support (Quick Install)
+
+If you **don't need PDF processing** functionality, you can skip system dependencies:
+
+```r
+# Install without optional dependencies
+install.packages("chatAI4R", dependencies = FALSE)
+
+# Load the package
+library(chatAI4R)
+```
+
+**What works:**
+- ✅ All AI chat functions (`chat4R`, `gemini4R`, etc.)
+- ✅ Multi-LLM execution (`multiLLMviaionet`)
+- ✅ Text processing and summarization
+- ✅ R package development tools
+- ✅ All core, workflow, and expertise functions
+
+**What doesn't work:**
+- ❌ `chatAI4pdf()` - PDF document analysis
+
+The package will show a helpful error message if you try to use PDF functions without `pdftools` installed.
+
+---
+
+## 🔍 Troubleshooting
+
+### Issue: "libqpdf.so: cannot open shared object file"
+
+**Solution:** The system library is installed but not found. Update the library cache:
+
+```bash
+sudo ldconfig
+```
+
+Then restart R and try installing pdftools again.
+
+### Issue: "Permission denied" during system library installation
+
+**Solution:** You need administrator privileges. Use `sudo` with the installation commands.
+
+### Issue: "Package not found" when installing system libraries
+
+**Solution:** Update your package manager's repository list first:
+
+**Ubuntu/Debian:**
+```bash
+sudo apt-get update
+```
+
+**CentOS/RHEL:**
+```bash
+sudo yum update
[…]
