### 2025-10-28 eb91b6f
Remove chatAI4pdf() from main package and create standalone script
- Remove chatAI4pdf.R function file and documentation
- Remove pdftools from DESCRIPTION Suggests dependencies
- Remove chatAI4pdf export from NAMESPACE
- Create chatAI4pdf_standalone.sh script for separate PDF functionality
- Update README.md to reflect standalone script usage
- Update test files to reference standalone script

This resolves installation issues caused by pdftools/qpdf dependency
while preserving PDF functionality as an optional standalone component.

Resolves #24

🤖 Generated with [Claude Code](https://claude.
[…]
