+++ chatAI4pdf_standalone.sh
@@ -1,0 +1,148 @@
+#!/bin/bash
+
+# chatAI4pdf Standalone Script
+# This script provides the chatAI4pdf functionality as a separate component
+# that can be sourced independently of the main chatAI4R package.
+# 
+# Usage:
+#   source chatAI4pdf_standalone.sh
+#   R -e "chatAI4pdf('path/to/file.pdf')"
+#
+# Requirements:
+#   - R with chatAI4R package installed (excluding chatAI4pdf function)
+#   - pdftools R package
+#   - qpdf system library
+#
+# Installation of system dependencies:
+#   Ubuntu/Debian: sudo apt-get install libqpdf-dev
+#   CentOS/RHEL:   sudo yum install qpdf-devel  
+#   Rocky Linux:   sudo dnf install qpdf-devel
+#   macOS:         brew install qpdf
+
+echo "Loading chatAI4pdf standalone functionality..."
+
+# Check if R is available
+if ! command -v R &> /dev/null; then
+    echo "Error: R is not installed or not in PATH"
+    exit 1
+fi
+
+# Check if qpdf system library is available
+if ! command -v qpdf &> /dev/null; then
+    echo "Warning: qpdf command not found. The pdftools R package may not work correctly."
+    echo "Please install qpdf system library:"
+    echo "  Ubuntu/Debian: sudo apt-get install libqpdf-dev"
+    echo "  CentOS/RHEL:   sudo yum install qpdf-devel"
+    echo "  Rocky Linux:   sudo dnf install qpdf-devel"
+    echo "  macOS:         brew install qpdf"
+fi
+
+# Create the chatAI4pdf function in R
+R --slave --no-restore --no-save << 'EOF'
+
+# Define chatAI4pdf function
+chatAI4pdf <- function(pdf_file_path,
+                       nch = 2000,
+                       verbose = TRUE) {
+
+  # Check if chatAI4R is installed
+  if (!requireNamespace("chatAI4R", quietly = TRUE)) {
+    stop(
+      "Package 'chatAI4R' is required but is not installed.\n",
+      "Please install it with: install.packages('chatAI4R')",
+      call. = FALSE
+    )
+  }
+
+  # Check if pdftools is installed
+  if (!requireNamespace("pdftools", quietly = TRUE)) {
+    stop(
+      "Package 'pdftools' is required for PDF processing but is not installed.\n",
+      "Please install it with: install.packages('pdftools')\n\n",
+      "Note: On Linux, you may need to install system dependencies first:\n",
+      "  Ubuntu/Debian: sudo apt-get install libqpdf-dev\n",
+      "  CentOS/RHEL:   sudo yum install qpdf-devel\n",
+      "  Rocky Linux:   sudo dnf install qpdf-devel\n",
+      "  macOS:         brew install qpdf\n",
+      "See the standalone script documentation for more information.",
+      call. = FALSE
+    )
+  }
+
+  # Load required packages
+  library(chatAI4R)
+  
+  #selection
+  choices1 <- c("All text",  "First Quarter Text",  "Second Quarter Text",
+                "Third Quarter Text", "Fourth Quarter Text")
+  selection1 <- utils::menu(choices1, title = "Where would you summarize the text?")
+
+  if (selection1 == 1) {
+    pds <- c(0, 1)
+  } else if (selection1 == 2) {
+    pds <- c(0, 0.25)
+  } else if (selection1 == 3) {
+    pds <- c(0.25, 0.5)
+  } else if (selection1 == 4) {
+    pds <- c(0.5, 0.75)
[…]
