+++ README.md
@@ -78,7 +78,7 @@ sudo dnf install qpdf-devel
 brew install qpdf
 ```
 
-**Note:** The `qpdf` system library is required for the `pdftools` dependency, which enables PDF processing functionality in `chatAI4pdf()`.
+**Note:** PDF processing functionality has been moved to a standalone script. See `chatAI4pdf_standalone.sh` for PDF analysis capabilities that require the `qpdf` system library.
 
 ### 2. Start R / RStudio console.
 
@@ -293,7 +293,7 @@ Expert-level functions that provide sophisticated data analysis, pattern recogni
 |convertScientificLiterature|Convert text to scientific literature format|[Script](https://github.com/kumeS/chatAI4R/blob/main/R/convertScientificLiterature.R)|
 |summaryWebScrapingText|Web scraping with intelligent summarization|[Script](https://github.com/kumeS/chatAI4R/blob/main/R/summaryWebScrapingText.R)|
 |geminiGrounding4R|Advanced AI with Google Search grounding|[Script](https://github.com/kumeS/chatAI4R/blob/main/R/geminiGrounding4R.R)|
-|chatAI4pdf|Intelligent PDF document analysis and summarization|[Script](https://github.com/kumeS/chatAI4R/blob/main/R/chatAI4pdf.R)|
+|chatAI4pdf|Intelligent PDF document analysis and summarization|[Standalone Script](https://github.com/kumeS/chatAI4R/blob/main/chatAI4pdf_standalone.sh)|
 |textFileInput4ai|Large-scale text file analysis with chunking|[Script](https://github.com/kumeS/chatAI4R/blob/main/R/textFileInput4ai.R)|
 |searchFunction|Expert-level R function discovery and recommendation|[Script](https://github.com/kumeS/chatAI4R/blob/main/R/searchFunction.R)|
 
