### 2025-10-28 156c2be
refactor: remove pdftools dependency and chatAI4pdf function to fix Linux installation
Addresses Linux installation error: "dependency 'pdftools' is not available for package 'chatAI4R'"

Changes:
- Removed pdftools from DESCRIPTION Suggests (line 36)
- Removed chatAI4pdf() function and documentation
- Cleaned up README.md installation instructions (removed PDF-specific steps)
- Removed chatAI4pdf from function table in README.md
- Moved chatAI4pdf_standalone.sh to inst/sh/ directory for optional use
- Deleted obsolete LINUX_INSTALL.md documentation
- Updated package version to 1.1.0

Linux us
[…]
