+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: chatAI4R
 Type: Package
 Title: Chat-Based Interactive Artificial Intelligence for R
-Version: 1.0.0
-Date: 2025-07-01
+Version: 1.1.0
+Date: 2025-10-29
 Authors@R: c(
     person (given = "Satoshi", family = "Kume",
     role = c("aut", "cre"), 
