### 2025-10-28 2a8d269
ci: add multi-platform R CMD check GitHub Actions workflow
Added comprehensive CI/CD testing across Windows, macOS, and Linux platforms.

Features:
- Tests on ubuntu-latest (R-release and R-devel)
- Tests on macos-latest (R-release)
- Tests on windows-latest (R-release)
- Automated R CMD check on push/PR to main, master, dev branches
- Manual workflow dispatch support
- Package dependency caching for faster builds
- Linux system dependencies installation (libcurl, libharfbuzz, libfribidi)
- Upload check results on failure for debugging

Badge:
- Added R-CMD-check status badge to README.md

Thi
[…]
