### 2025-10-28 0e96384
docs: update README.md for version 1.1.0 release
Updated documentation to reflect v1.1.0 changes and improvements.

Changes:
- Updated version badge from 1.0.0 to 1.1.0
- Added comprehensive v1.1.0 changelog section
  - Linux installation fix (pdftools dependency removal)
  - CI/CD integration with GitHub Actions
  - Multi-platform testing (Windows/macOS/Linux)
  - Installation improvements and simplified setup
- Reorganized "What's New" section with clear version separators
- Updated function count (54 → 53 functions)
- Updated citation version to 1.1.0
- Updated BibTeX entry with correct ver
[…]
