### 2025-10-28 kumeS/chatAI4R#24 @kumeS
@khoa-yelo Now it is ready to install on your linux system.
