### 2025-10-29 kumeS/chatAI4R#24 @kumeS
@khoa-yelo
Thank you for your feedback.
could you share your install command on R?
the latest version of installation is:
devtools::install_github("kumeS/chatAI4R")
library(chatAI4R)
