### 2025-10-29 kumeS/chatAI4R#24 @khoa-yelo
I'm testing installation through CRAN
```
install.packages("chatAI4R")
library(chatAI4R)
```
If you rebuild CRAN, I can test it one last time. Regardless, one of the guideline I receive from JOSS is to check `Whether a potential user can easily install, understand, and test the software`
As R has lots of systems dependency and a user may not have permission to change OS libraries - this is my scenario which is also typical many users on HPC of an organization. Having a container for your application would really accelerate my testing process and make sure other users can adopt your tool easily. Thanks
