### 2025-10-29 773a23f
test: add comprehensive test runner with detailed reporting
Added new test framework for systematic testing of all chatAI4R functions.

Features:
- Comprehensive test runner (run_all_tests.R)
  - Tests all functions without stopping on errors
  - Smart API key detection and test skipping
  - Detailed error tracking with function names and messages
  - Performance metrics for each test
  - Summary statistics (passed/failed/skipped counts)
  - Top 5 slowest tests reporting
  - Saves results to Rds file for analysis

- Test categories:
  - Utility functions (no API required): 3 tests
  - Core API
[…]
