+++ tests/test/run_all_tests.R
@@ -1,0 +1,438 @@
+# =============================================================================
+# Comprehensive Test Runner for chatAI4R Package
+# =============================================================================
+#
+# Purpose: Test all chatAI4R functions systematically
+# Features:
+#   - Runs all function tests without stopping on errors
+#   - Tracks success/failure for each function
+#   - Provides detailed error messages for failed tests
+#   - Generates comprehensive summary report
+#
+# Usage:
+#   source("tests/test/run_all_tests.R")
+#
+# Requirements:
+#   - API keys must be set (optional, tests will skip if not available)
+#   - RStudio not required (clipboard tests will be skipped)
+#
+# =============================================================================
+
+# Initialize test tracking
+test_results <- list()
+test_start_time <- Sys.time()
+
+# Function to safely run a test and capture results
+run_test_safely <- function(test_name, test_code, skip_reason = NULL) {
+  if (!is.null(skip_reason)) {
+    cat(sprintf("[SKIP] %s - %s\n", test_name, skip_reason))
+    return(list(
+      function_name = test_name,
+      status = "skipped",
+      reason = skip_reason,
+      error = NULL,
+      duration = 0
+    ))
+  }
+
+  cat(sprintf("[TEST] Running: %s\n", test_name))
+  start_time <- Sys.time()
+
+  result <- tryCatch({
+    eval(parse(text = test_code))
+    end_time <- Sys.time()
+    duration <- as.numeric(difftime(end_time, start_time, units = "secs"))
+
+    cat(sprintf("[PASS] %s (%.2f sec)\n", test_name, duration))
+    list(
+      function_name = test_name,
+      status = "passed",
+      error = NULL,
+      duration = duration
+    )
+  }, error = function(e) {
+    end_time <- Sys.time()
+    duration <- as.numeric(difftime(end_time, start_time, units = "secs"))
+
+    cat(sprintf("[FAIL] %s - %s (%.2f sec)\n", test_name, e$message, duration))
+    list(
+      function_name = test_name,
+      status = "failed",
+      error = e$message,
+      duration = duration
+    )
+  }, warning = function(w) {
+    end_time <- Sys.time()
+    duration <- as.numeric(difftime(end_time, start_time, units = "secs"))
+
+    cat(sprintf("[WARN] %s - %s (%.2f sec)\n", test_name, w$message, duration))
+    list(
+      function_name = test_name,
+      status = "warning",
+      error = w$message,
+      duration = duration
+    )
+  })
+
+  return(result)
+}
+
+# Check if API keys are available
+has_openai <- nzchar(Sys.getenv("OPENAI_API_KEY"))
+has_gemini <- nzchar(Sys.getenv("GoogleGemini_API_KEY"))
+has_replicate <- nzchar(Sys.getenv("Replicate_API_KEY"))
+has_ionet <- nzchar(Sys.getenv("IONET_API_KEY"))
+has_rstudio <- !is.null(getOption("rstudio.version"))
+
+cat("=============================================================================\n")
+cat("chatAI4R Package - Comprehensive Test Suite\n")
+cat("=============================================================================\n\n")
+cat("Environment Check:\n")
+cat(sprintf("  OpenAI API Key: %s\n", ifelse(has_openai, "✓ Available", "✗ Not set")))
+cat(sprintf("  Gemini API Key: %s\n", ifelse(has_gemini, "✓ Available", "✗ Not set")))
+cat(sprintf("  Replicate API Key: %s\n", ifelse(has_replicate, "✓ Available", "✗ Not set")))
[…]
