+++ tests/test/README.md
@@ -1,0 +1,292 @@
+# chatAI4R Test Suite
+
+Comprehensive testing framework for the chatAI4R package.
+
+## Overview
+
+This test suite provides automated testing for all chatAI4R functions with detailed reporting and error tracking.
+
+## Files
+
+- **`run_all_tests.R`**: Main test runner script with comprehensive reporting
+- **`all_examples_combined.R`**: Original combined examples file (reference)
+- **`test_results.Rds`**: Saved test results (generated after running tests)
+
+## Features
+
+✅ **Comprehensive Testing**: Tests all available functions systematically
+✅ **Error Resilience**: Continues execution even when tests fail
+✅ **Detailed Reporting**: Shows exactly which tests passed/failed and why
+✅ **Performance Metrics**: Tracks execution time for each test
+✅ **Smart Skipping**: Automatically skips tests that require unavailable API keys or RStudio
+✅ **Summary Statistics**: Provides clear pass/fail rates and counts
+
+## Usage
+
+### Basic Usage
+
+```r
+# From R console
+source("tests/test/run_all_tests.R")
+```
+
+### From Command Line
+
+```bash
+# Run tests from command line
+Rscript tests/test/run_all_tests.R
+
+# Run tests and save output
+Rscript tests/test/run_all_tests.R > test_output.txt 2>&1
+```
+
+### With CI/CD
+
+```yaml
+# GitHub Actions example
+- name: Run chatAI4R tests
+  run: |
+    Rscript tests/test/run_all_tests.R
+```
+
+## Requirements
+
+### Required
+
+- R (>= 4.2.0)
+- chatAI4R package installed
+
+### Optional (for full test coverage)
+
+API Keys (set as environment variables):
+
+```r
+Sys.setenv(OPENAI_API_KEY = "your-openai-key")
+Sys.setenv(GoogleGemini_API_KEY = "your-gemini-key")
+Sys.setenv(Replicate_API_KEY = "your-replicate-key")
+Sys.setenv(IONET_API_KEY = "your-ionet-key")
+```
+
+Or set in `.Renviron`:
+
+```bash
+OPENAI_API_KEY=your-openai-key
+GoogleGemini_API_KEY=your-gemini-key
+Replicate_API_KEY=your-replicate-key
+IONET_API_KEY=your-ionet-key
+```
+
+## Output Format
+
+### Console Output
+
+```
+=============================================================================
+chatAI4R Package - Comprehensive Test Suite
+=============================================================================
+
+Environment Check:
+  OpenAI API Key: ✓ Available
+  Gemini API Key: ✗ Not set
+  Replicate API Key: ✗ Not set
+  io.net API Key: ✗ Not set
+  RStudio: ✗ Not available
+
+[INIT] Loading chatAI4R package...
+[INIT] Package loaded successfully
+
+=============================================================================
+Starting Test Execution
+=============================================================================
+
+[TEST] Running: 01. ngsub
+[PASS] 01. ngsub (0.01 sec)
+
+[TEST] Running: 02. removeQuotations
+[PASS] 02. removeQuotations (0.00 sec)
+
+[TEST] Running: 03. chat4R
+[PASS] 03. chat4R (1.23 sec)
+
+[TEST] Running: 04. textEmbedding
+[FAIL] 04. textEmbedding - API rate limit exceeded (0.45 sec)
+
+[SKIP] 05. gemini4R - Gemini API key not set
+
+=============================================================================
+Test Execution Complete
+=============================================================================
+
+Summary Statistics:
+===================
+Total Tests:     15
+Passed:          10 (66.7%)
[…]
