### 2025-10-29 b6853b1
refactor: consolidate test suite and remove duplicate test files
Streamlined test directory structure by removing obsolete test files.

Changes:
- Removed duplicate/obsolete test files:
  - all_examples_combined.R (replaced by run_all_tests.R)
  - all_examples_combined_v2.R (partial examples, obsolete)
  - test_execution.R (v1.0.0 test script, replaced)
  - run_tests.sh (shell wrapper, replaced by R script)

- Completely rewrote tests/test/README.md:
  - Clear quick start section
  - Detailed feature documentation
  - Test category descriptions (14 tests)
  - Example output format
  - Comprehe
[…]
