+++ tests/test/run_tests.sh
@@ -1,333 +1,0 @@
-#!/bin/bash
-
-# =============================================================================
-# chatAI4R Package Test Runner v1.0.0
-# =============================================================================
-# Comprehensive testing tool for chatAI4R package with multi-API support
-# Tests 25+ functions across utility, API, and advanced functionality
-# 
-# Features:
-# • No-API utility testing for quick validation
-# • OpenAI API integration testing (chat4R, textEmbedding, vision4R, etc.)
-# • Multi-LLM testing via io.net API (NEW in v1.0.0)
-# • Extended API testing (Gemini, Replicate, Dify)
-# • Comprehensive error handling and logging
-# 
-# Usage: ./run_tests.sh --help
-# =============================================================================
-
-set -e  # Exit on any error
-
-# Colors for output
-RED='\033[0;31m'
-GREEN='\033[0;32m'
-YELLOW='\033[1;33m'
-BLUE='\033[0;34m'
-NC='\033[0m' # No Color
-
-# Default values
-API_KEY=""
-GEMINI_KEY=""
-REPLICATE_KEY=""
-DIFY_KEY=""
-DEEPL_KEY=""
-IONET_KEY=""
-MODE="utilities"
-VERBOSE=true
-
-# Help function
-show_help() {
-    echo "=========================================="
-    echo "  chatAI4R Package Test Runner v1.0.0"
-    echo "=========================================="
-    echo ""
-    echo "DESCRIPTION:"
-    echo "  Comprehensive testing tool for chatAI4R package with multi-API support."
-    echo "  Tests 25+ functions across utility, API, and advanced functionality."
-    echo ""
-    echo "USAGE:"
-    echo "  $0 [OPTIONS]"
-    echo ""
-    echo "API KEY OPTIONS:"
-    echo "  -k, --api-key KEY      🔑 OpenAI API key (sk-...)"
-    echo "                         └── Required for: chat4R, textEmbedding, vision4R, etc."
-    echo "  -g, --gemini-key KEY   🔑 Google Gemini API key"  
-    echo "                         └── Required for: gemini4R, geminiGrounding4R"
-    echo "  -r, --replicate-key KEY 🔑 Replicate API token"
-    echo "                         └── Required for: replicatellmAPI4R (Llama models)"
-    echo "  -d, --dify-key KEY     🔑 Dify API key"
-    echo "                         └── Required for: DifyChat4R (workflow AI)"
-    echo "  -l, --deepl-key KEY    🔑 DeepL API key"
-    echo "                         └── Required for: discussion flow translation"
-    echo "  -i, --ionet-key KEY    🔑 io.net API key (NEW in v1.0.0)"
-    echo "                         └── Required for: multiLLMviaionet, multiLLM_random5/10"
-    echo ""
-    echo "TEST MODE OPTIONS:"
-    echo "  -m, --mode MODE        Test execution mode:"
-    echo "    utilities            📦 Utility functions only (no API key needed)"
-    echo "                         └── Tests: ngsub, removeQuotations, slow_print_v2"
-    echo "    api-only             🌐 Basic API functions only (OpenAI key required)" 
-    echo "                         └── Tests: chat4R, textEmbedding, vision4R, etc."
-    echo "    full                 🚀 All utility + API functions (recommended)"
[…]
