+++ tests/test/all_examples_combined.R
@@ -1,550 +1,0 @@
-# =============================================================================
-# Combined Examples from chatAI4R Package
-# Generated on: 2025-07-01 20:49:23.741767
-# =============================================================================
-
-# Load the chatAI4R package
-library(chatAI4R)
-
-# Note: Some examples may require API keys to be set:
-# Sys.setenv(OPENAI_API_KEY = 'your_openai_key')
-# Sys.setenv(GoogleGemini_API_KEY = 'your_gemini_key')
-# Sys.setenv(Replicate_API_KEY = 'your_replicate_key')
-# Sys.setenv(IONET_API_KEY = 'your_ionet_key')
-
-# =============================================================================
-
-# -----------------------------------------------------------------------------
-# addCommentCode examples
-# -----------------------------------------------------------------------------
-
-# Option 1
-# Select some text in RStudio and then run the rstudio addins
-# Option 2
-# Copy the text into your clipboard then execute
-addCommentCode(Model = "gpt-4o-mini", language = "English", SelectedCode = TRUE)
-
-# -----------------------------------------------------------------------------
-# addRoxygenDescription examples
-# -----------------------------------------------------------------------------
-
-# Option 1
-# Select some text in RStudio and then run the rstudio addins
-# Option 2
-# Copy the text into your clipboard then execute
-addRoxygenDescription(Model = "gpt-4o-mini", SelectedCode = FALSE)
-
-# -----------------------------------------------------------------------------
-# autocreateFunction4R examples
-# -----------------------------------------------------------------------------
-
-#Sys.setenv(OPENAI_API_KEY = "<APIKEY>")
-autocreateFunction4R(Func_description = "2*n+3 sequence")
-
-# -----------------------------------------------------------------------------
-# chat4R_history examples
-# -----------------------------------------------------------------------------
-
-#Sys.setenv(OPENAI_API_KEY = "Your API key")
-history <- list(list('role' = 'system', 'content' = 'You are a helpful assistant.'),
-list('role' = 'user', 'content' = 'Who won the world series in 2020?'))
-chat4R_history(history)
-
-# -----------------------------------------------------------------------------
-# chat4R_streaming examples
-# -----------------------------------------------------------------------------
-
-#Sys.setenv(OPENAI_API_KEY = "Your API key")
-# Without system_set
-chat4R_streaming(content = "What is the capital of France?")
-# With system_set provided
-chat4R_streaming(
-content = "What is the capital of France?",
-system_set = "You are a helpful assistant."
-)
-
-# -----------------------------------------------------------------------------
-# chat4R examples
-# -----------------------------------------------------------------------------
-
-#Sys.setenv(OPENAI_API_KEY = "Your API key")
-chat4R(content = "What is the capital of France?", check = TRUE)
-chat4R(content = "What is the capital of France?", check = FALSE)
-
-# -----------------------------------------------------------------------------
-# chat4Rv2 examples
-# -----------------------------------------------------------------------------
-
-#Sys.setenv(OPENAI_API_KEY = "Your API key")
-# Using chat4Rv2 without system_prompt (default behavior)
-response <- chat4Rv2(content = "What is the capital of France?")
-response
-# Using chat4Rv2 with a system_prompt provided
-response <- chat4Rv2(content = "What is the capital of France?",
-                     system_prompt = "You are a helpful assistant.")
-response
-
-# -----------------------------------------------------------------------------
-# chatAI4pdf examples - MOVED TO STANDALONE SCRIPT
-# -----------------------------------------------------------------------------
-# Note: chatAI4pdf functionality has been moved to chatAI4pdf_standalone.sh
-# Use: source('chatAI4pdf_standalone.sh') to access this functionality
-
[…]
