### 2025-10-31 kumeS/chatAI4R#27 @tzuV
[JOSS Review] Streamlining of language in the turorials
Can the tutorials be updated to be in one language (preferably in English)? It could greatly benefit the clarity and consistency.

Some of the videos are very helpful, like: https://www.youtube.com/watch?v=VJaltAS9Ef8

However, it becomes hard to follow if the model does as expected when the output is in a different language than the input language.

Also, the following video: https://www.youtube.com/watch?v=wTMZfLMXYH0 seems to be a demonstration of chatGPT's capabilites. But it would be more appropriate to see this in the context of the functionality of the package.

Written material instead of videos would also be beneficial.
