### 2025-10-31 #9283 @tzuV
@kumeS could you look into this issue regarding the enhancement of the tutorial?
https://github.com/kumeS/chatAI4R/issues/27
