### 2025-10-31 kumeS/chatAI4R#29 @tzuV
[JOSS Review] autocreateFunction4R does not create functions with gpt4-nano
I have tried the function ` autocreateFunction4R`.

I acknowledge that the function is experimental.

When I run the function, I would expect it to create an R function based on the text I provide in the Func_description parameter. However, it produces a lot of text, which seems to be a bit random:

```

> autocreateFunction4R(Func_description = "2*n+3 sequence")
01 Run function creation
In your function/ the function should accept- a 2-dimension with n rows and n columns A matrices: A1 – A2n- a 3D matrix B1: [n-1,n,dim] which represents n-1 rows is class n, where row n is empty.- Result text format should be “n copies of r print B1, and then the computation will be identical to your function. example: [r1] B1 for row 3, [r2] B1 for row 4, …[r5] B1 for row 6.”- n-1 handles row where B1 is class N.Result should be- Empty matrix, row is integer 2n+3-1 between [0, 2n

ProjectShare_Supervisor is a Python application project to create an interactive Python REST API on top of MySQL. Python REST API and MySQL are two open-source project codes. Therefore, ProjectShare_Supervisor, a project to create Python REST API and MySQL are open-source codes. Python REST API creates a harmonicial communication with client systems02 Propose improvements to the function
To be published in Wikibooks – CS188 – Artificial Intelligence – Professor Patrick Henry Winston Harvard University, Fall 2016.03 Improve the function  that and license model suitable for our project. Then, we invite research to publish results in journals or conferences. The outcome expected from this project is published reviewed open-project.
 that wants to take benefit from this interactive Python REST API. This contract and agreement focus in presenting open solutions to other researches rather than our ownership of the project; therefore, it is an port project with Harvard License Manger- before identifying hacker- if not, ProjectShare-Coordinator will notify hacker to take legal action and will with-hold the reward by Harvard, the licensing company. Then, legal actions on internet, researched prayers will protect innocent. The Harvard Rights Exceptions has heavy sanction against hacking, such as heavy jail imprisonment and heavy fines. Oracle had tried for 212 million US dollars, and lost $ about $46 million.@ 2016-09-15;
04 Include roxygen comments
 Source Code Readme ID = pDev - Software Development License; Programmer comments only.Comment: The aim of roxygen comments is instructing potential hackers about the architecture of the current roxygen comments in this single functions and about sharing better roxygen comments.Please use roxygen in comments. Suppose that roxygen comments is not possible without an extra roxygenDoc() means. The purpose of roxygen is to make more attractive comments comments. Even if we had to describe for tomorrow why we need them, we may be a parent of tomorrow and we hope that the future generations to gather together and rely on IT, since Genie fiber optics strategy may alter the electricity became obsolete within 1 Genie through future generations.Please keep it in mansiter all. com. Every public_APIDoc_xxx_common function should be able to send back the roxygen result files to mansiter. Thomas-E-Boehm. org in a convenient format. Then, his build.com or build\.com website can also provide a anonymous- roxygen-comment-spam-hunt.org IFNAME website that incorporates the API documentation string for searching comments under under sub-directory. An .RWill, .RCode, .RDoxyFile can be dropped in
 that wants to take benefit from this interactive Python REST API. This contract and agreement focus in presenting open solutions to other researches rather than our ownership of the project; therefore, it is an port project with Harvard License Manger- before identifying hacker- if not, ProjectShare-Coordinator will notify hacker to take legal action and will with-hold the reward by Harvard, the licensing company. Then, legal actions on internet, researched prayers will protect innocent. The Harvard Rights Exceptions has heavy sanction against hacking, such as heavy jail imprisonment and heavy fines. Oracle had tried for 212 million US dollars, and lost $ about $46 million.@ 2016-09-15;  Finished!![1] " Source Code Readme ID = pDev - Software Development License; Programmer comments only.Comment: The aim of roxygen comments is instructing potential hackers about the architecture of the current roxygen comments in this single functions and about sharing better roxygen comments.Please use roxygen in comments. Suppose that roxygen comments is not possible without an extra roxygenDoc() means. The purpose of roxygen is to make more attractive comments comments. Even if we had to describe for tomorrow why we need them, we may be a parent of tomorrow and we hope that the future generations to gather together and rely on IT, since Genie fiber optics strategy may alter the electricity became obsolete within 1 Genie through future generations.Please keep it in mansiter all. com. Every public_APIDoc_xxx_common function should be able to send back the roxygen result files to mansiter. Thomas-E-Boehm. org in a convenient format. Then, his build.com or build\\.com website can also provide a anonymous- roxygen-comment-spam-hunt.org IFNAME website that incorporates the API documentation string for searching comments under under sub-directory. An .RWill, .RCode, .RDoxyFile can be dropped in\n that wants to take benefit from this interactive Python REST API. This contract and agreement focus in presenting open solutions to other researches rather than our ownership of the project; therefore, it is an port project with Harvard License Manger- before identifying hacker- if not, ProjectShare-Coordinator will notify hacker to take legal action and will with-hold the reward by Harvard, the licensing company. Then, legal actions on internet, researched prayers will protect innocent. The Harvard Rights Exceptions has heavy sanction against hacking, such as heavy jail imprisonment and heavy fines. Oracle had tried for 212 million US dollars, and lost $ about $46 million.@ 2016-09-15;  "
````
