### 2025-10-31 #9283 @tzuV
@kumeS There is some issues which the autocreateFunction4R function. I do acknowledge it to be experimental.
https://github.com/kumeS/chatAI4R/issues/29
