### 2025-10-31 kumeS/chatAI4R#30 @tzuV
[JOSS Review] chat4R is missing the argument "check"
In the file tests/test/all_examples_combined.R, there is the following example:

```{r}
# -----------------------------------------------------------------------------
# chat4R examples
# -----------------------------------------------------------------------------

#Sys.setenv(OPENAI_API_KEY = "Your API key")
chat4R(content = "What is the capital of France?", check = TRUE)
chat4R(content = "What is the capital of France?", check = FALSE)
```
However, it would seem that the argument has been removed from the function?
