### 2025-10-31 #9283 @tzuV
@kumeS There is some inconsistency between the documentation and a test file. More details can be found here:
https://github.com/kumeS/chatAI4R/issues/30
