### 2025-10-31 kumeS/chatAI4BI#32 @tzuV
[JOSS Review] createEBAYdes() BUG
It seems as the function does not work when verbose is set to true.

I get the following when it is set to true:

```
> # -----------------------------------------------------------------------------
> # create_eBay_Description examples
> # -----------------------------------------------------------------------------
>
> # Option 1
> # Select some text in RStudio and then run the rstudio addins
> # Option 2
> # Copy the text into your clipboard then execute
> clipr::write_clip("A new A100 GPU")
>
> createEBAYdes(Model = "gpt-4o-mini", SelectedCode = FALSE, verbose = TRUE)

 createEBAYdes:
  |================================================================================================================================| 100%
Error in strsplit(text, "") : non-character argument
```

However, when I set verbose to false, it seems it works properly:

```
> # -----------------------------------------------------------------------------
> # create_eBay_Description examples
> # -----------------------------------------------------------------------------
>
> # Option 1
> # Select some text in RStudio and then run the rstudio addins
> # Option 2
> # Copy the text into your clipboard then execute
> clipr::write_clip("A new A100 GPU")
>
> createEBAYdes(Model = "gpt-4o-mini", SelectedCode = FALSE, verbose = FALSE)
>
> clipr::read_clip()
 [1] "**Product Title: New A100 GPU – Unleash Your Computing Power with Free Shipping!**"
 [2] ""
 [3] "---"
 [4] ""
 [5] "**Product Description:**"
 [6] ""
 [7] "Elevate your computing experience with the all-new A100 GPU! Designed specifically for high-performance computing, deep learning, and data analytics, this powerhouse graphics processing unit is engineered to handle the most demanding workloads with unparalleled efficiency. Whether you're a gamer, a data scientist, or a creative professional, the A100 GPU is your ticket to extraordinary performance and speed."
 [8] ""
 [9] "**Features:**"
[10] ""
[11] "- **Revolutionary Architecture:** The A100 GPU is built on NVIDIA's Ampere architecture, optimizing core performance and allowing for faster computations."
[12] "- **High Memory Capacity:** Equipped with 40 GB of high-bandwidth memory (HBM2) to handle massive datasets."
[13] "- **Multi-Instance GPU Technology:** Run multiple applications simultaneously without compromising on performance."
[14] "- **AI Acceleration:** Seamlessly accelerate AI workloads for enhanced performance and faster time-to-insight."
[15] "- **Versatile Compatibility:** Installs easily into various systems, enhancing both desktop and server performance."
[16] ""
[17] "---"
[18] ""
[19] "**Included Accessories/Contents:**"
[20] ""
[21] "- A100 GPU"
[22] "- User Manual"
[23] "- Low-profile bracket for custom configurations"
[24] "- Installation CD/Software"
[25] ""
[26] "---"
[27] ""
[28] "**Product Specifications:**"
[29] ""
[30] "- **Year of Release:** 2020"
[31] "- **Material:** Aluminum chassis with advanced thermal technology"
[32] "- **Size:** "
[33] "  - Length: 26.7 cm (10.5 inches)"
[34] "  - Width: 12.5 cm (4.9 inches)"
[35] "  - Height: 4.5 cm (1.8 inches)"
[36] "- **Weight:** 2.9 kg (6.4 lbs)"
[37] "- **Memory Capacity:** 40 GB HBM2"
[38] "- **Languages listed:** English, Spanish, German, French (multi-language support)"
[39] ""
[40] "---"
[41] ""
[42] "**Friendly Comment:**"
[43] ""
[44] "Transform your computing game with the A100 GPU! Don’t miss out on the chance to make your projects shine brighter and run faster—grab yours today and watch your potential soar! 🚀 "
[45] ""
[46] "**Buy now and enjoy FREE SHIPPING to the United States!** "
[47] ""
[48] "---"
[49] ""
[50] "Optimize your listings and reach the audience that needs finely-tuned performance. Good luck with your eBay success!\r"
```
