### 2025-10-31 #9283 @tzuV
@kumeS i found a bug in the createEBAYdes.
https://github.com/kumeS/chatAI4R/issues/32
