### 2025-10-31 kumeS/chatAI4R#34 @tzuV
Bug in createRfunction (verbose)
Similar to https://github.com/kumeS/chatAI4R/issues/32 i get an error when verbose is set to true.

Here is a snippet of the console with verbose set to false and true:

```
> # -----------------------------------------------------------------------------
> # createRfunction examples
> # -----------------------------------------------------------------------------
>
> #Copy the idea text of the R function to your clipboard and run this function.
> clipr::write_clip("A function which can compute the mean of a vector of any size")
>
> createRfunction(SelectedCode = FALSE, verbose = FALSE, SlowTone = TRUE)
>
> clipr::read_clip()
 [1] "mean_vector <- function(vec) {"               "  # Check if the input is a numeric vector"
 [3] "  if (!is.numeric(vec)) {"                    "    stop(\"Input must be a numeric vector\")"
 [5] "  }"                                          "  "
 [7] "  # Compute the mean of the vector"           "  mean_value <- sum(vec) / length(vec)"
 [9] "  "                                           "  # Return the computed mean"
[11] "  return(mean_value)"                         "}\r"
>
> # -----------------------------------------------------------------------------
> # createRfunction examples
> # -----------------------------------------------------------------------------
>
> #Copy the idea text of the R function to your clipboard and run this function.
> clipr::write_clip("A function which can compute the mean of a vector of any size")
>
> createRfunction(SelectedCode = FALSE, verbose = FALSE, SlowTone = FALSE)
>
> clipr::read_clip()
 [1] "mean_vector <- function(vec) {"                "  # Check if the input is a numeric vector"
 [3] "  if (!is.numeric(vec)) {"                     "    stop(\"Input must be a numeric vector.\")"
 [5] "  }"                                           "  "
 [7] "  # Calculate the mean of the vector"          "  result <- sum(vec) / length(vec)"
 [9] "  "                                            "  # Return the computed mean"
[11] "  return(result)"                              "}\r"
>
> # -----------------------------------------------------------------------------
> # createRfunction examples
> # -----------------------------------------------------------------------------
>
> #Copy the idea text of the R function to your clipboard and run this function.
> clipr::write_clip("A function which can compute the mean of a vector of any size")
>
> createRfunction(SelectedCode = FALSE, verbose = TRUE, SlowTone = FALSE)

 createRfunction:
  |================================================================================================================================| 100%
Error in strsplit(text, "") : non-character argument
```
