### 2025-10-31 #9283 @tzuV
@kumeS i have found a bug in the createRfunction
https://github.com/kumeS/chatAI4R/issues/34
