### 2025-10-31 kumeS/chatAI4R#27 @kumeS
@tzuV Thank you for your suggestions.

I will update the tutorial video of "chatAI4R package demo in R" (https://www.youtube.com/watch?v=VJaltAS9Ef8) with English literature.

Regarding the video [Roxygen2 (GPT-4 UI) / LLM roxygen demo], I have currently developed the addRoxygenDescription function as the same feature using the UI.

I will remove this video URL and add a description for this functionality.
