### 2025-11-01 kumeS/chatAI4R#38 @tzuV
[Question] Some functions cannot be called
The following functions cannot be called from the package:

- gemini4R
- geminiGrounding4R
- interpretResult
- multiLLMviaionet

I can see that the functions are in the source code folder and in the NAMESPACE file. But when i try to use them or find them in the package ``` chatAI4R ``` it is not present?

My installation of the package have been with the github installation: ``` devtools::install_github("kumeS/chatAI4R") ```

Please let me know if i am missing something.
