### 2025-11-01 #9283 @tzuV
@kumeS i have some problems using some certain functions in the package. Please see more in the issue. https://github.com/kumeS/chatAI4R/issues/38
