### 2025-11-03 kumeS/chatAI4BI#40 @khoa-yelo
[JOSS Review] - Documentation
A couple points on documentation
- Who is the target audience for your software?
- Could you include a written tutorial to demonstrate the use of your software in a real world application to show the key point you mention in the paper that `chatAI4R` is not simply an API wrapper, etc...?

As of now, the tool feels like a collection of lots of functions but I'm not sure its target audience and how they could use it in their project.
