### 2025-11-04 kumeS/chatAI4R#43 @tzuV
[JOSS REVIEW] - summaryWebScrapingText bug
I have tried to run the example for the function with verbose set to TRUE and FALSE, but i get an empty list every time:

```
> # -----------------------------------------------------------------------------
> # summaryWebScraping examples
> # -----------------------------------------------------------------------------
>
> summaryWebScrapingText(query = "LLM", t = "w", gl = "us", hl = "en", URL_num = 5, verbose = TRUE)
list()
```

The concept is cool, but it is very fragile in many ways. One thing is that webpages are dynamic, so scraping demands a lot of maintenance.

I wonder if the function is beyond the scope of this package? I think this function could be removed from this package and perhaps be added to another project (perhaps about using llm's for data gathering and processing?).
