### 2025-11-04 #9283 @tzuV
@kumeS I don't have much more to add in terms of the functionalities. I have some remaining points in regard to the writing of the project paper and checking on the opportunity for contributions. As far as I remember, it seemed fine, so I don't expect too much more on that front. If you address the issues I have provided, I think it will conclude my review

Generally, I have had a lot of fun going through, and I see the value in most of the functions. I agree with @khoa-yelo in regard to https://github.com/kumeS/chatAI4R/issues/40, and I wondered if some of the functions were even necessary for the package.

I'll return later with a pull request for the .Rmd for the vignette where I have tested all the functions. You can decide if you want it. It is not a requirement for me to have it included in the package.
