+++ CONTRIBUTING.md
@@ -31,7 +31,7 @@ Using chatAI4R for a paper you are writing? Consider [citing it][citation].
 
 Using chatAI4R and got stuck? Browse the [documentation][website] to see if you can find a solution. Still stuck? Post your question as an [issue on GitHub][new_issue]. While we cannot offer user support, we'll try to do our best to address it, as questions often lead to better documentation or the discovery of bugs.
 
-Want to ask a question in private? Contact the package maintainer by [email][mailto:j.skelmose@rn.dk].
+Want to ask a question in private? Contact the package maintainer by [email][mailto: satoshi.kume@a.riken.jp].
 
 ### Propose an idea 💡
 
