### 2025-11-05 kumeS/chatAI4R#43 @kumeS
Thank you for your feedback.
I will review your comments as
soon as possible.
Best regards, Satoshi
