### 2025-11-05 kumeS/chatAI4R#43 @kumeS
Thank you for your comment. I will remove the function of summaryWebScrapingText from this package.
