### 2025-11-05 c244d49
v1.x: JOSS version
