+++ tests/test/phase3_tests.R
@@ -1,0 +1,227 @@
+# =============================================================================
+# Phase 3: Coverage Improvement (Target: 80%)
+# =============================================================================
+#
+# Adding 9 functions to reach 42/52 (80% coverage)
+# Current: 33/52 (63.5%) → Target: 42/52 (80%)
+#
+# Focus:
+#   1. Interactive functions - existence tests
+#   2. RStudio-dependent functions - existence tests
+#   3. Remaining API functions
+#
+# =============================================================================
+
+#' Run Phase 3 Tests
+#'
+#' @param verbose Logical, print detailed output
+#' @return Test results
+#' @export
+run_phase3_tests <- function(verbose = TRUE) {
+  
+  has_openai <- nzchar(Sys.getenv("OPENAI_API_KEY"))
+  has_gemini <- nzchar(Sys.getenv("GoogleGemini_API_KEY"))
+  
+  test_results <- list()
+  test_start_time <- Sys.time()
+  
+  run_test_safely <- function(test_name, test_code, skip_reason = NULL) {
+    if (!is.null(skip_reason)) {
+      if (verbose) cat(sprintf("[SKIP] %s - %s\n", test_name, skip_reason))
+      return(list(function_name = test_name, status = "skipped", reason = skip_reason, error = NULL, duration = 0))
+    }
+    
+    if (verbose) cat(sprintf("[TEST] Running: %s\n", test_name))
+    start_time <- Sys.time()
+    
+    result <- tryCatch({
+      eval(parse(text = test_code))
+      end_time <- Sys.time()
+      duration <- as.numeric(difftime(end_time, start_time, units = "secs"))
+      if (verbose) cat(sprintf("[PASS] %s (%.2f sec)\n", test_name, duration))
+      list(function_name = test_name, status = "passed", error = NULL, duration = duration)
+    }, error = function(e) {
+      end_time <- Sys.time()
+      duration <- as.numeric(difftime(end_time, start_time, units = "secs"))
+      if (verbose) cat(sprintf("[FAIL] %s - %s (%.2f sec)\n", test_name, e$message, duration))
+      list(function_name = test_name, status = "failed", error = e$message, duration = duration)
+    })
+    
+    return(result)
+  }
+  
+  if (verbose) {
+    cat("═══════════════════════════════════════════════════════════════════════════\n")
+    cat("Phase 3 Test Suite - Coverage Improvement to 80%\n")
+    cat("═══════════════════════════════════════════════════════════════════════════\n\n")
+  }
+  
+  suppressPackageStartupMessages(library(chatAI4R))
+  
+  test_idx <- 1
+  
+  # ═════════════════════════════════════════════════════════════════════════
+  # Group 1: Interactive Functions - Existence Tests
+  # ═════════════════════════════════════════════════════════════════════════
+  
+  # Test 1: TextSummary
+  test_results[[test_idx]] <- run_test_safely(
+    sprintf("%02d. TextSummary", test_idx),
+    'f <- get("TextSummary")
+     stopifnot(is.function(f))
+     stopifnot(length(formals(f)) > 0)'
+  )
+  test_idx <- test_idx + 1
+  
+  # Test 2: TextSummaryAsBullet
+  test_results[[test_idx]] <- run_test_safely(
+    sprintf("%02d. TextSummaryAsBullet", test_idx),
+    'f <- get("TextSummaryAsBullet")
+     stopifnot(is.function(f))
+     stopifnot(length(formals(f)) > 0)'
+  )
+  test_idx <- test_idx + 1
+  
+  # Test 3: revisedText
[…]
