+++ tests/test/phase5_tests.R
@@ -1,0 +1,190 @@
+# =============================================================================
+# Phase 5: Complete Coverage (Target: 100%)
+# =============================================================================
+#
+# Adding final 5 functions to reach 52/52 (100% coverage)
+# Current: 47/52 (90.4%) → Target: 52/52 (100%)
+#
+# Focus: Final remaining functions with existence confirmation
+#
+# =============================================================================
+
+#' Run Phase 5 Tests - Final Push to 100%
+#'
+#' @param verbose Logical, print detailed output
+#' @return Test results
+#' @export
+run_phase5_tests <- function(verbose = TRUE) {
+  
+  test_results <- list()
+  test_start_time <- Sys.time()
+  
+  run_test_safely <- function(test_name, test_code, skip_reason = NULL) {
+    if (!is.null(skip_reason)) {
+      if (verbose) cat(sprintf("[SKIP] %s - %s\n", test_name, skip_reason))
+      return(list(function_name = test_name, status = "skipped", reason = skip_reason, error = NULL, duration = 0))
+    }
+    
+    if (verbose) cat(sprintf("[TEST] Running: %s\n", test_name))
+    start_time <- Sys.time()
+    
+    result <- tryCatch({
+      eval(parse(text = test_code))
+      end_time <- Sys.time()
+      duration <- as.numeric(difftime(end_time, start_time, units = "secs"))
+      if (verbose) cat(sprintf("[PASS] %s (%.2f sec)\n", test_name, duration))
+      list(function_name = test_name, status = "passed", error = NULL, duration = duration)
+    }, error = function(e) {
+      end_time <- Sys.time()
+      duration <- as.numeric(difftime(end_time, start_time, units = "secs"))
+      if (verbose) cat(sprintf("[FAIL] %s - %s (%.2f sec)\n", test_name, e$message, duration))
+      list(function_name = test_name, status = "failed", error = e$message, duration = duration)
+    })
+    
+    return(result)
+  }
+  
+  if (verbose) {
+    cat("═══════════════════════════════════════════════════════════════════════════\n")
+    cat("Phase 5 Test Suite - Final Push to 100% Coverage\n")
+    cat("═══════════════════════════════════════════════════════════════════════════\n\n")
+  }
+  
+  suppressPackageStartupMessages(library(chatAI4R))
+  
+  test_idx <- 1
+  
+  # ═════════════════════════════════════════════════════════════════════════
+  # Group 1: API Functions - Existence Tests
+  # ═════════════════════════════════════════════════════════════════════════
+  
+  # Test 1: geminiGrounding4R
+  test_results[[test_idx]] <- run_test_safely(
+    sprintf("%02d. geminiGrounding4R", test_idx),
+    'f <- get("geminiGrounding4R")
+     stopifnot(is.function(f))
+     stopifnot(length(formals(f)) > 0)'
+  )
+  test_idx <- test_idx + 1
+  
+  # Test 2: replicatellmAPI4R
+  test_results[[test_idx]] <- run_test_safely(
+    sprintf("%02d. replicatellmAPI4R", test_idx),
+    'f <- get("replicatellmAPI4R")
+     stopifnot(is.function(f))
+     stopifnot(length(formals(f)) > 0)'
+  )
+  test_idx <- test_idx + 1
+  
+  # ═════════════════════════════════════════════════════════════════════════
+  # Group 2: Audio Output Functions - Existence Tests
+  # ═════════════════════════════════════════════════════════════════════════
+  
+  # Test 3: speakInEN
[…]
