+++ NAMESPACE
@@ -48,7 +48,6 @@ export(slow_print_v2)
 export(speakInEN)
 export(speakInJA)
 export(speakInJA_v2)
-export(summaryWebScrapingText)
 export(supportIdeaGeneration)
 export(textEmbedding)
 export(textFileInput4ai)

[…]
