### 2025-11-05 4e61b4c
v1.x: JOSS version
