### 2025-11-13 a9824e2
v1.x: JOSS version
