+++ NAMESPACE
@@ -96,15 +96,9 @@ importFrom(parallel,stopCluster)
 importFrom(rstudioapi,getActiveDocumentContext)
 importFrom(rstudioapi,insertText)
 importFrom(rstudioapi,isAvailable)
-importFrom(rvest,html_attr)
-importFrom(rvest,html_nodes)
-importFrom(rvest,html_text)
-importFrom(rvest,read_html)
 importFrom(stats,runif)
-importFrom(utils,URLencode)
 importFrom(utils,askYesNo)
 importFrom(utils,capture.output)
 importFrom(utils,menu)
 importFrom(utils,setTxtProgressBar)
 importFrom(utils,txtProgressBar)
-importFrom(xml2,read_html)
