### 2025-11-14 kumeS/chatAI4R#30 @kumeS
The file of tests/test/all_examples_combined.R has been removed when previously updating the package.
I have tried chat4R function with the argument "check = TRUE" and it's option is working well.

```r
> chat4R(content = "What is the capital of France?", check = TRUE)
Status: No error
                          content
1 The capital of France is Paris.
```
