+++ .github/workflows/R-CMD-check.yml
@@ -58,7 +58,7 @@ jobs:
         if: runner.os == 'Linux'
         run: |
           sudo apt-get update
-          sudo apt-get install -y libcurl4-openssl-dev libharfbuzz-dev libfribidi-dev
+          sudo apt-get install -y libcurl4-openssl-dev libharfbuzz-dev libfribidi-dev libglpk-dev libxml2-dev libgmp-dev
 
       - name: Install dependencies
         run: |
