### 2025-11-14 0229b8e
v1.x: JOSS version
