+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: chatAI4R
 Type: Package
 Title: Chat-Based Interactive Artificial Intelligence for R
-Version: 1.1.2
+Version: 1.1.3
 Date: 2025-11-14
 Authors@R: c(
     person (given = "Satoshi", family = "Kume",
