### 2025-11-14 kumeS/chatAI4R#32 @kumeS
Hi @tzuV, in my R environment, I was able to execute this function

The error "Error in strsplit(text, "") : non-character argument" is not a bug in the function, but is likely due to the randomness of the AI's output.

Therefore, I have strengthened the error handling of the createEBAYdes function & slow_print_v2 function to cope with fluctuations in AI output.

Updates:  https://github.com/kumeS/chatAI4R/commit/0229b8ef792aa5c8477d9915c539a657bc85c03c

Here are the execution results in my R environment.

```r
> clipr::write_clip("A new A100 GPU")
> createEBAYdes(Model = "gpt-4o-mini", SelectedCode = FALSE, verbose = TRUE)

 createEBAYdes:
  |=================================================================| 100%
**Product Title: Brand New NVIDIA A100 GPU - Unleash the Power of AI and Data Science!**

---

**Product Description:**

Introducing the NVIDIA A100 GPU, a cutting-edge powerhouse designed to handle the most demanding tasks in artificial intelligence, machine learning, and data science. With unparalleled performance and efficiency, the A100 is essential for researchers, developers, and technology enthusiasts looking to elevate their projects to the next level.

Harnessing the power of the NVIDIA Ampere architecture, the A100 provides outstanding performance for both training and inference in AI applications, making it a must-have for anyone serious about leveraging technology to solve complex problems.

---

**Features:**

- **Next-generation NVIDIA Ampere architecture** for unrivaled performance and efficiency.
- **Multi-instance GPU (MIG) capabilities** allow you to partition the GPU to maximize utilization and performance across multiple workloads.
- **Support for both FP32 and Tensor Float 32 (TF32)** to optimize both training and inferencing processes.
- **NVLink and PCIe Gen 4** support for seamless integration with other GPUs and high-speed data transfer.
- Enhanced **AI and machine learning capabilities**, optimized memory bandwidth for large datasets.

---

**Included Accessories/Contents:**

- NVIDIA A100 GPU
- Quick Start Guide
- Warranty Information
- Installation Support (available upon request)

---

**Product Specifications:**

- **Year of Release:** 2020
- **Material:** High-performance semiconductor material (Silicon)
- **Size:**
  - Dimensions: 26.67 cm x 11.4 cm x 4.4 cm (10.5 in x 4.49 in x 1.73 in)
- **Weight:** 2.4 kg (5.29 lbs)
- **Memory Capacity:** 40 GB or 80 GB HBM2 (configurable)
- **Supported Languages:** English (documentation), with additional language support in the software.

---

**Encouraging Comment:**

Don't miss your chance to supercharge your AI projects with the NVIDIA A100 GPU! With features designed to tackle the most intensive workloads, this GPU is the key to unlocking your next great discovery. Get yours today and watch your performance soar!

**Free Shipping to the United States!**
```
