### 2025-11-14 55e3773
v1.x: JOSS version
