### 2025-11-15 50c438b
v1.x: JOSS version
