### 2025-11-15 8734414
v1.x: JOSS version
