### 2025-11-15 c50b4c4
v1.x: JOSS version
