### 2025-11-15 7fe411c
v1.x: JOSS version
