### 2025-11-15 bb8f450
v1.x: JOSS version
