### 2025-11-15 ccdbbd4
v1.x: JOSS version
