### 2025-11-15 f4642bf
v1.x: JOSS version
