### 2025-11-15 8ffb530
v1.x: JOSS version
