### 2025-11-16 ad1af7d
v1.1.x: JOSS review
