### 2025-11-16 kumeS/chatAI4R#40 @kumeS
@khoa-yelo  Thank you for the feedback on documentation clarity.

## Target Audience

**Primary**: R package developers (original focus)
- Automated Roxygen2 documentation, function generation, code quality improvement

**Secondary**: Data analysts & researchers (expanded functionality)
- Statistical interpretation, literature processing, knowledge extraction

The chatAI4R project started as a development support tool for R packages, but it is being developed into a data analysis support tool by integrating functions such as data analysis interpretation and knowledge extraction, as well as multi-LLM support.

## Beyond Simple API Wrapper

chatAI4R implements a **4-layer architecture**:
- Layer 1: Basic API access
- Layer 2: Conversation memory, context management
- Layer 3: Multi-bot systems, R package automation
- Layer 4: Domain-specific analysis (13+ domains)

As described in the paper, I believe that this package enables progressively sophisticated AI-powered workflows beyond simple API calls.

## Documentation Updates : https://github.com/kumeS/chatAI4R/commit/6d3ba48bc92be5ee9b1bec28eec5e9f3ce243c9e

Based on your feedback, we've updated the README with:
- Clear primary/secondary audience identification
- Quick Start examples demonstrating the layered architecture
- Simplified use cases

As a quick start of README.md, I've presented some examples.
If you give me a little more time, I'd like to add more examples.

I had a similar comment from tzuV at https://github.com/kumeS/chatAI4R/issues/27.
Video demonstration [R / LLM demo (2) create R function](https://youtu.be/dIUkCqCS0Tw) shows the R package development workflow in RStudio. However, a apart of this video is in Japanese so I plan to create English-language tutorials later.
