### 2025-11-16 6d3ba48
v1.1.x: JOSS review
