+++ README.md
@@ -19,41 +19,37 @@
 
 chatAI4R is an experimental project aimed at developing and implementing various LLM applications in R. Furthermore, the package is under continuous development with a focus on extending its capabilities for bioinformatics analysis.
 
-## About this project and future developments
+## 🎯 Who is this for?
 
-- **🚀 Multi-API AI Integration with R** *(v1.0.0 Features)*
-  - [OpenAI API](https://platform.openai.com/docs/api-reference/introduction) (ChatGPT, GPT-4, text embeddings, vision) ✅ **Enhanced Security**
-  - [Google Gemini API](https://ai.google.dev/) (Gemini models, search grounding) ✅ **Enhanced Security**
-  - [Replicate API](https://replicate.com/) (Llama, other open-source models) ✅ **Enhanced Security**
-  - [Dify API](https://dify.ai/) (Workflow-based AI applications) ✅ **Enhanced Security**
-  - [**NEW**: io.net API](https://io.net/) (Multi-LLM parallel execution, **23+ models**)
-  - DeepL API (Professional translation)
+- R package developers - automate documentation (Roxygen2), generate functions, improve code quality
+- Data analysts - interpret statistical results, process literature, extract insights
 
-- **🔒 Security Enhancements (v1.0.0)**
-  - **HTTP Status Validation**: All API functions now validate HTTP response codes
-  - **Safe Data Access**: Null-safe nested JSON parsing across all functions  
-  - **Enhanced Error Handling**: Comprehensive error messages without credential exposure
-  - **Input Validation**: Standardized parameter validation using assertthat patterns
+**Key Difference**: Not just an API wrapper - provides a **4-layer architecture** from basic API access to expert-level workflows and domain-specific analysis tools.
 
-- **4-Layer Architecture for Progressive AI Capabilities**
-  - **Core Layer**: Direct API access and basic utilities
-  - **Usage/Task Layer**: Conversation management, text processing, proofreading
-  - **Workflow Layer**: Multi-bot systems, R package development automation
-  - **Expertise Layer**: Advanced data mining, pattern recognition, expert analysis
+## ✨ Key Features
 
-- **LLM-assisted R Package Development**
-  - Complete package design and architecture planning
-  - Automated R function creation with documentation (✨ **Updated**: `autocreateFunction4R` now uses modern `chat4R` API)
-  - Code optimization and error analysis
-  - Professional text proofreading and enhancement
+### 🚀 Multi-API Support (6 AI Services)
+- **OpenAI** (ChatGPT, GPT-4, Vision, Embeddings)
+- **Google Gemini** (Gemini models with search grounding)
+- **io.net** (23+ models including DeepSeek-R1, Llama-4, Qwen3, Mistral)
+- **Replicate** (Open-source models)
+- **Dify** (Workflow-based AI)
+- **DeepL** (Professional translation)
 
-- **Advanced Data Analysis and Knowledge Discovery**
-  - Multi-domain statistical analysis interpretation (13+ domains)
-  - Scientific literature processing and knowledge extraction
-  - Web data mining and intelligent summarization
-  - Expert-level discussion simulation and peer review processes
-  
-***The functionality for interlanguage translation using DeepL has been separated as the 'deepRstudio' package. Functions related to text-to-image generation were separated as the 'stableDiffusion4R' package.***
[…]
