### 2025-11-17 015b69f
v1.1.x: JOSS review
