+++ man/replicatellmAPI4R.Rd
@@ -9,6 +9,7 @@ replicatellmAPI4R(
   model_url,
   simple = TRUE,
   fetch_stream = FALSE,
+  collapse_tokens = TRUE,
   api_key = Sys.getenv("Replicate_API_KEY")
 )
 }
@@ -21,13 +22,17 @@ replicatellmAPI4R(
 
 \item{fetch_stream}{A logical value indicating whether to fetch a streaming response. Default is FALSE.}
 
+\item{collapse_tokens}{A logical value indicating whether to collapse token vectors into a single string for LLM text outputs. When TRUE (default), token vectors like c("The", " capital", " of", " France") are automatically combined into "The capital of France". This parameter only affects LLM text outputs and has no effect on other model types (e.g., image URLs, audio data). Default is TRUE.}
+
 \item{api_key}{A character string representing the Replicate API key. Defaults to the environment variable "Replicate_API_KEY".}
 }
 \value{
-If fetch_stream is FALSE, returns either a simplified output (if simple is TRUE) or the full API response. In streaming mode, outputs the response stream directly to the console.
+If fetch_stream is FALSE, returns either a simplified output (if simple is TRUE) or the full API response. When simple is TRUE and collapse_tokens is TRUE, returns a single character string. In streaming mode, outputs the response stream directly to the console.
 }
 \description{
 This function interacts with the Replicate API (v1) to utilize language models (LLM) such as Llama. It sends a POST request with the provided input and handles both streaming and non-streaming responses.
+
+**Note**: This function is primarily designed and optimized for Large Language Models (LLMs). While the Replicate API supports various model types (e.g., image generation, audio models), this function's features (especially `collapse_tokens`) are tailored for LLM text outputs. For non-LLM models, consider setting `collapse_tokens = FALSE` or use the full response mode with `simple = FALSE`.
 }
 \examples{
 \dontrun{
@@ -46,8 +51,16 @@ This function interacts with the Replicate API (v1) to utilize language models (
     )
   )
   model_url <- "/models/meta/meta-llama-3.1-405b-instruct/predictions"
-  response <- replicatellmAPI4R(input, model_url)
+
+  # Default: collapse_tokens = TRUE (returns single string)
+  response <- replicatellmAPI4R(input, model_url, fetch_stream = TRUE)
   print(response)
+  # [1] "The capital of France is Paris."
+
+  # Keep tokens separated
+  response_tokens <- replicatellmAPI4R(input, model_url, fetch_stream = TRUE, collapse_tokens = FALSE)
+  print(response_tokens)
+  # [1] "The" " capital" " of" " France" " is" " Paris" "."
 }
 
 }
