### 2025-11-17 fd4a20f
v1.1.x: JOSS review
