+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: chatAI4R
 Type: Package
 Title: Chat-Based Interactive Artificial Intelligence for R
-Version: 1.1.7
+Version: 1.1.8
 Date: 2025-11-18
 Authors@R: c(
     person (given = "Satoshi", family = "Kume",
