### 2025-11-22 kumeS/chatAI4R#27 @tzuV
@kumeS Thank you for the updated documentation. It gives better clarity for new users to use the package.

I can see this video (https://www.youtube.com/watch?v=wTMZfLMXYH0) is still in the readme.Rmd. When the url has been removed, we can close this issue.
