### 2025-11-22 kumeS/chatAI4R#29 @tzuV
I just tested the function again and it works! I like how it tries to improve the function iteratively. Good work @kumeS
