### 2025-11-24 kumeS/chatAI4R#38 @tzuV
I tried running the 'geminiGrounding4R' with the example code:
```{r}
# -----------------------------------------------------------------------------
# geminiGrounding4R examples
# -----------------------------------------------------------------------------

# Synchronous text generation with grounding:
result <- geminiGrounding4R(
  mode = "text",
  contents = "What is the current Google stock price?",
  store_history = FALSE,
  dynamic_threshold = 1,
  api_key = Sys.getenv("GoogleGemini_API_KEY")
)
print(result)
```

But I get an error:
`
Error: HTTP error: 400
Response: {
  "error": {
    "code": 400,
    "message": "Invalid JSON payload received. Unknown name \"googleSearchRetrieval\" at 'tools[0]': Proto field is not repeating, cannot start list.",
    "status": "INVALID_ARGUMENT",
    "details": [
      {
        "@type": "type.googleapis.com/google.rpc.BadRequest",
        "fieldViolations": [
          {
            "field": "tools[0]",
            "description": "Invalid JSON payload received. Unknown name \"googleSearchRetrieval\" at 'tools[0]': Proto field is not repeating, cannot start list."
          }
        ]
      }
    ]
  }
}
`
