### 2025-11-27 #9283 @kumeS
Hi @khoa-yelo, thank you for your comment.
I will finalize all issues and comments this weekend.
All the best,
Satoshi
