### 2025-11-30 4aab316
v1.1.x: JOSS review
