### 2025-11-30 6c13ef7
v1.1.x: JOSS review
