### 2025-11-30 kumeS/chatAI4R#40 @kumeS
I have updated the paper.md (https://github.com/kumeS/chatAI4R/commit/6c13ef7d4162e1a8e2688f9273826f9c6e1c999f).
