+++ R/geminiGrounding4R.R
@@ -1,4 +1,4 @@
-#' Gemini API Google Search Grounding Request (v1beta, 2025-07)
+#' Gemini API Google Search Grounding Request (v1/v1beta, 2025-07)
 #'
 #' A thin R wrapper for Google Gemini API with Google Search grounding enabled.
 #' Grounding improves factuality by incorporating web search results.
@@ -6,10 +6,11 @@
 #' @param mode        One of `"text"`, `"stream_text"`, `"chat"`, `"stream_chat"`.
 #' @param contents    Character vector (single-turn) or list of message objects
 #'                    (chat modes). See Examples.
-#' @param model       Gemini model ID. Default `"gemini-2.0-flash"`.
+#' @param model       Gemini model ID. Default `"gemini-2.5-flash"`.
+#' @param api_version API version for the Generative Language API (`"v1"` or `"v1beta"`). Default `"v1beta"`.
 #' @param store_history Logical. If TRUE, chat history is persisted to the
 #'                      `chat_history` env-var (JSON).
-#' @param dynamic_threshold Numeric [0,1] for dynamic retrieval threshold (default: 0.7).
+#' @param dynamic_threshold Numeric [0,1] for dynamic retrieval threshold (default: 0.7). Only used for Gemini 1.5 models with `googleSearchRetrieval`; ignored for newer models that use `google_search`.
 #' @param api_key     Your Google Gemini API key (default: `Sys.getenv("GoogleGemini_API_KEY")`).
 #' @param max_tokens  Maximum output tokens. Default 2048.
 #' @param debug       Logical. If TRUE, prints request details for debugging.
@@ -29,7 +30,6 @@
 #'     mode = "text",
 #'     contents = "What is the current Google stock price?",
 #'     store_history = FALSE,
-#'     dynamic_threshold = 0.7,
 #'     debug = TRUE,  # Enable debug to see request details
 #'     api_key = Sys.getenv("GoogleGemini_API_KEY")
 #'   )
@@ -72,7 +72,8 @@
 
 geminiGrounding4R <- function(mode,
                                contents,
-                               model             = "gemini-1.5-pro",
+                               model             = "gemini-2.5-flash",
+                               api_version       = "v1beta",
                                store_history     = FALSE,
                                dynamic_threshold = 0.7,
                                api_key           = Sys.getenv("GoogleGemini_API_KEY"),
@@ -91,14 +92,19 @@ geminiGrounding4R <- function(mode,
   if (!mode %in% valid_modes)
     stop("Invalid mode. Choose one of: ", paste(valid_modes, collapse = ", "), call. = FALSE)
 
+  api_version <- tolower(api_version)
+  if (!api_version %in% c("v1", "v1beta")) {
+    stop('api_version must be "v1" or "v1beta".', call. = FALSE)
+  }
+
   #----- Build endpoint --------------------------------------------------------
   method   <- if (grepl("^stream", mode)) "streamGenerateContent" else "generateContent"
   endpoint <- sprintf(
-    "https://generativelanguage.googleapis.com/v1beta/models/%s:%s?key=%s",
-    model, method, api_key
+    "https://generativelanguage.googleapis.com/%s/models/%s:%s",
+    api_version, model, method
   )
   if (grepl("^stream", mode))
-    endpoint <- paste0(endpoint, "&alt=sse")  # SSE required for streaming
+    endpoint <- paste0(endpoint, "?alt=sse")  # SSE required for streaming
 
   #----- Construct request body ------------------------------------------------
   make_message <- function(role, text) {
[…]
