### 2025-11-30 kumeS/chatAI4R#38 @kumeS
Updated: https://github.com/kumeS/chatAI4R/commit/b6d964fdf35090a350c2e2dc3996eed18c5b1078

I have updated the Google API endpoints and modified the geminiGrounding4R function.

Now it's function works.

```r
> result <- geminiGrounding4R(
+   mode = "text",
+   contents = "What is the current Google stock price?",
+   store_history = FALSE,
+   debug = TRUE,
+   api_key = Sys.getenv("GoogleGemini_API_KEY")
+ )
dynamic_threshold is ignored for google_search on this model.
=== DEBUG: Request Body ===
{"contents":[{"role":"user","parts":[{"text":"What is the current Google stock price?"}]}],"tools":[{"google_search":{}}],"generationConfig":{"maxOutputTokens":2048}}
=== DEBUG: Endpoint ===
https://generativelanguage.googleapis.com/v1beta/models/gemini-2.5-flash:generateContent
========================
> print(result)
$candidates
$candidates[[1]]
$candidates[[1]]$content
$candidates[[1]]$content$parts
$candidates[[1]]$content$parts[[1]]
$candidates[[1]]$content$parts[[1]]$text
[1] "As of November 29, 2025, the stock prices for Alphabet (Google) are as follows:\n\n*   **Alphabet Inc. Class A (GOOGL)** shares are valued at $320.57.\n*   **Alphabet Inc. Class C (GOOG)** shares are trading at approximately $320.12 to $320.80."

$candidates[[1]]$content$role
[1] "model"

$candidates[[1]]$finishReason
[1] "STOP"

$candidates[[1]]$index
[1] 0

$candidates[[1]]$groundingMetadata
$candidates[[1]]$groundingMetadata$searchEntryPoint
$candidates[[1]]$groundingMetadata$searchEntryPoint$renderedContent
[1] "<style>\n.container {\n  align-items: center;\n  border-radius: 8px;\n  display: flex;\n  font-family: Google Sans, Roboto, sans-serif;\n  font-size: 14px;\n  line-height: 20px;\n  padding: 8px 12px;\n}\n.chip {\n  display: inline-block;\n  border: solid 1px;\n  border-radius: 16px;\n  min-width: 14px;\n  padding: 5px 16px;\n  text-align: center;\n  user-select: none;\n  margin: 0 8px;\n  -webkit-tap-highlight-color: transparent;\n}\n.carousel {\n  overflow: auto;\n  scrollbar-width: none;\n  white-space: nowrap;\n  margin-right: -12px;\n}\n.headline {\n  display: flex;\n  margin-right: 4px;\n}\n.gradient-container {\n  position: relative;\n}\n.gradient {\n  position: absolute;\n  transform: translate(3px, -9px);\n  height: 36px;\n  width: 9px;\n}\n@media (prefers-color-scheme: light) {\n  .container {\n    background-color: #fafafa;\n    box-shadow: 0 0 0 1px #0000000f;\n  }\n  .headline-label {\n    color: #1f1f1f;\n  }\n  .chip {\n    background-color: #ffffff;\n    border-color: #d2d2d2;\n    color: #5e5e5e;\n    text-decoration: none;\n  }\n  .chip:hover {\n    background-color: #f2f2f2;\n  }\n  .chip:focus {\n    background-color: #f2f2f2;\n  }\n  .chip:active {\n    background-color: #d8d8d8;\n    border-color: #b6b6b6;\n  }\n  .logo-dark {\n    display: none;\n  }\n  .gradient {\n    background: linear-gradient(90deg, #fafafa 15%, #fafafa00 100%);\n  }\n}\n@media (prefers-color-scheme: dark) {\n  .container {\n    background-color: #1f1f1f;\n    box-shadow: 0 0 0 1px #ffffff26;\n  }\n  .headline-label {\n    color: #fff;\n  }\n  .chip {\n    background-color: #2c2c2c;\n    border-color: #3c4043;\n    color: #fff;\n    text-decoration: none;\n  }\n  .chip:hover {\n    background-color: #353536;\n  }\n  .chip:focus {\n    background-color: #353536;\n  }\n  .chip:active {\n    background-color: #464849;\n    border-color: #53575b;\n  }\n  .logo-light {\n    display: none;\n  }\n  .gradient {\n    background: linear-gradient(90deg, #1f1f1f 15%, #1f1f1f00 10 ... <truncated>

$candidates[[1]]$groundingMetadata$groundingChunks
$candidates[[1]]$groundingMetadata$groundingChunks[[1]]
$candidates[[1]]$groundingMetadata$groundingChunks[[1]]$web
$candidates[[1]]$groundingMetadata$groundingChunks[[1]]$web$uri
[1] "https://vertexaisearch.cloud.google.com/grounding-api-redirect/AUZIYQHUrd_9QQFSvoFv3EhCmGE-MevlHg5m8llebeZTZtNtrShr6cD-9WvM0OeA-CUNGFqASf3XfP3NXOOWYg3XQ7R2hsaRLX6_cK3PFnpSAtV4LwuExBHl6z0MKXRyefmq5xJYEUWl"

$candidates[[1]]$groundingMetadata$groundingChunks[[1]]$web$title
[1] "robinhood.com"

$candidates[[1]]$groundingMetadata$groundingChunks[[2]]
$candidates[[1]]$groundingMetadata$groundingChunks[[2]]$web
$candidates[[1]]$groundingMetadata$groundingChunks[[2]]$web$uri
[1] "https://vertexaisearch.cloud.google.com/grounding-api-redirect/AUZIYQHeUsa5Qr_DJ1CjV03zYnEDtz0_zxIOcLbz2kZ6EfjXy_rnC0LdK-9RcexAf6U2P04NvPxPJUWyakdA82tYBYY9WqFmNG9LzF4H-1JKqj9-MY2jCa6jLGJSrByq4V5oARaFR3DoLU767fv7GA=="

$candidates[[1]]$groundingMetadata$groundingChunks[[2]]$web$title
[1] "tradingview.com"

$candidates[[1]]$groundingMetadata$groundingChunks[[3]]
$candidates[[1]]$groundingMetadata$groundingChunks[[3]]$web
$candidates[[1]]$groundingMetadata$groundingChunks[[3]]$web$uri
[1] "https://vertexaisearch.cloud.google.com/grounding-api-redirect/AUZIYQGSq_FT5MxT4OQwg-2AJb69yviHCfcOGJXQKDh7Gnas3EsJZzNEJQlETRSp2bCI9zZzUujfMmpwOUjaPY5wlLWZ05ZK3mQehKweWb2Y0Ms512PtzE1O6AO_OXmI_G5bYjq46r8="

$candidates[[1]]$groundingMetadata$groundingChunks[[3]]$web$title
[1] "robinhood.com"

$candidates[[1]]$groundingMetadata$groundingSupports
$candidates[[1]]$groundingMetadata$groundingSupports[[1]]
$candidates[[1]]$groundingMetadata$groundingSupports[[1]]$segment
$candidates[[1]]$groundingMetadata$groundingSupports[[1]]$segment$startIndex
[1] 101

$candidates[[1]]$groundingMetadata$groundingSupports[[1]]$segment$endIndex
[1] 148

$candidates[[1]]$groundingMetadata$groundingSupports[[1]]$segment$text
[1] "Class A (GOOGL)** shares are valued at $320.57."

$candidates[[1]]$groundingMetadata$groundingSupports[[1]]$groundingChunkIndices
$candidates[[1]]$groundingMetadata$groundingSupports[[1]]$groundingChunkIndices[[1]]
[1] 0

$candidates[[1]]$groundingMetadata$groundingSupports[[2]]
$candidates[[1]]$groundingMetadata$groundingSupports[[2]]$segment
$candidates[[1]]$groundingMetadata$groundingSupports[[2]]$segment$startIndex
[1] 169

$candidates[[1]]$groundingMetadata$groundingSupports[[2]]$segment$endIndex
[1] 241

$candidates[[1]]$groundingMetadata$groundingSupports[[2]]$segment$text
[1] "Class C (GOOG)** shares are trading at approximately $320.12 to $320.80."

$candidates[[1]]$groundingMetadata$groundingSupports[[2]]$groundingChunkIndices
$candidates[[1]]$groundingMetadata$groundingSupports[[2]]$groundingChunkIndices[[1]]
[1] 1

$candidates[[1]]$groundingMetadata$groundingSupports[[2]]$groundingChunkIndices[[2]]
[1] 2

$candidates[[1]]$groundingMetadata$webSearchQueries
$candidates[[1]]$groundingMetadata$webSearchQueries[[1]]
[1] "current Google stock price"

$candidates[[1]]$groundingMetadata$webSearchQueries[[2]]
[1] "GOOGL stock price"

$usageMetadata
$usageMetadata$promptTokenCount
[1] 9

$usageMetadata$candidatesTokenCount
[1] 107

$usageMetadata$totalTokenCount
[1] 364

$usageMetadata$promptTokensDetails
$usageMetadata$promptTokensDetails[[1]]
$usageMetadata$promptTokensDetails[[1]]$modality
[1] "TEXT"

$usageMetadata$promptTokensDetails[[1]]$tokenCount
[1] 9

$usageMetadata$toolUsePromptTokenCount
[1] 76

$usageMetadata$toolUsePromptTokensDetails
$usageMetadata$toolUsePromptTokensDetails[[1]]
$usageMetadata$toolUsePromptTokensDetails[[1]]$modality
[1] "TEXT"

$usageMetadata$toolUsePromptTokensDetails[[1]]$tokenCount
[1] 76

$usageMetadata$thoughtsTokenCount
[1] 172

$modelVersion
[1] "gemini-2.5-flash"

$responseId
[1] "yd4racfuHpim1e8PxYa7uQk"
```
