### 2025-11-30 9b25880
v1.1.x: JOSS review
