### 2025-11-30 dea40b7
v1.1.x: JOSS review
