+++ R/multiLLMviaionet.R
@@ -1,34 +1,41 @@
 #' Multi-LLM via io.net API
 #'
 #' @title multiLLMviaionet: Execute ALL available LLM models simultaneously via io.net API
-#' @description This function automatically runs the same prompt across ALL currently available 
-#'   LLM models on the io.net API. It supports running all models in parallel with streaming 
+#' @description This function automatically runs the same prompt across ALL currently available
+#'   LLM models on the io.net API (by default it now selects one random model to reduce API load).
+#'   It supports running multiple models (configurable) in parallel with streaming
 #'   responses and comprehensive error handling.
-#'   
+#'
 #'   The function dynamically fetches the current list of available models from the io.net API
-#'   and executes ALL of them by default (unless limited by max_models parameter). Results are 
+#'   and executes a configurable subset (default 1 random model; set max_models or random_selection to control). Results are
 #'   cached for 1 hour to improve performance. If the API is unavailable, it falls back to a static model list.
-#'   
-#'   Supported model categories include:
-#'   - Meta Llama series (Llama-4-Maverick, Llama-3.3-70B, etc.)
-#'   - DeepSeek series (DeepSeek-R1, DeepSeek-R1-Distill variants)
-#'   - Qwen series (Qwen3-235B, QwQ-32B, Qwen2.5-Coder, etc.)
-#'   - Specialized models (AceMath-7B, watt-tool-70B, ReaderLM-v2)
-#'   - Microsoft Phi, Mistral, Google Gemma, and more
+#'
+#'   Typical models observed on io.net (30-Nov-2025) include:
+#'   - moonshotai/Kimi-K2-Thinking, moonshotai/Kimi-K2-Instruct-0905
+#'   - zai-org/GLM-4.6
+#'   - meta-llama/Llama-4-Maverick-17B-128E-Instruct-FP8, Llama-3.3-70B-Instruct, Llama-3.2-90B-Vision-Instruct
+#'   - Qwen/Qwen3-235B-A22B-Thinking-2507, Qwen3-Next-80B-A3B-Instruct, Qwen2.5-VL-32B-Instruct, Intel/Qwen3-Coder-480B-A35B-Instruct-int4-mixed-ar
+#'   - deepseek-ai/DeepSeek-R1-0528
+#'   - mistralai/Devstral-Small-2505, Magistral-Small-2506, Mistral-Large-Instruct-2411, Mistral-Nemo-Instruct-2407
+#'   - openai/gpt-oss-120b, openai/gpt-oss-20b
 #'
 #' @param prompt A string containing the input prompt to send to all selected models.
-#' @param max_models Integer specifying maximum number of models to run (1-50). Default is all available models.
+#' @param max_models Integer specifying maximum number of models to run (1-50). Default is 1 (single model).
 #' @param streaming Logical indicating whether to use streaming responses. Default is FALSE.
-#' @param random_selection Logical indicating whether to randomly select models from all available models. 
-#'   If TRUE, randomly selects up to max_models from all available models.
-#'   If FALSE, uses all available models in order (up to max_models limit). Default is FALSE.
-#' @param api_key A string containing the io.net API key. 
+#' @param random_selection Logical indicating whether to randomly select models from all available models.
+#'   If TRUE (default), randomly selects up to max_models from all available models.
+#'   If FALSE, uses models in order (up to max_models limit).
+#' @param api_key A string containing the io.net API key.
 #'   Defaults to the environment variable "IONET_API_KEY".
 #' @param max_tokens Integer specifying maximum tokens to generate per model. Default is 1024.
 #' @param temperature Numeric value controlling randomness (0-2). Default is 0.7.
[…]
