+++ man/print_multiLLM_results.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/multiLLMviaionet.R
+\name{print_multiLLM_results}
+\alias{print_multiLLM_results}
+\title{Pretty-print multiLLMviaionet results}
+\usage{
+print_multiLLM_results(result, show_failed = TRUE, trim = NULL)
+}
+\arguments{
+\item{result}{A result object returned by `multiLLMviaionet()`.}
+
+\item{show_failed}{Logical; if TRUE (default), also prints failed/timeout models and their errors.}
+
+\item{trim}{Optional integer; if not NULL, truncates long responses to the first `trim` characters for display.}
+}
+\value{
+Invisibly returns NULL.
+}
+\description{
+Display model responses with status, handling NULLs and errors cleanly.
+}
