+++ man/ionet_models.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/multiLLMviaionet.R
+\name{ionet_models}
+\alias{ionet_models}
+\title{ionet_models: Retrieve the available model list from io.net API}
+\usage{
+ionet_models(api_key = NULL, verbose = TRUE, force_refresh = FALSE)
+}
+\arguments{
+\item{api_key}{Optional API key for fetching current models. Defaults to IONET_API_KEY environment variable.}
+
+\item{verbose}{Logical indicating whether to show detailed fetching information. Default is TRUE.}
+
+\item{force_refresh}{Logical indicating whether to force refresh the model list from API. Default is FALSE.}
+}
+\value{
+A character vector of model names, or a data frame if detailed=TRUE in the underlying call.
+}
+\description{
+Convenience wrapper around list_ionet_models() to get the current model list.
+  By default it uses cached data; set force_refresh = TRUE to refetch.
+}
+\details{
+List io.net Models (simple alias)
+}
