### 2025-11-30 kumeS/chatAI4R#38 @kumeS
## About the multiLLMvioionet function

According to a person involved with io.net, this is less an implementation issue with the multiLLMvioionet function and more likely due to a recent API specification change that may have slowed execution responses for large models, including openai/gpt-oss-120b.

On the package/R function side, I made fixes such as enhancing error handling and reducing the number of models executed.

I also added a function, print_multiLLM_results, to output the results.

```r
> # Basic usage - single random model (default)
> result_one <- multiLLMviaionet(
+   prompt = "Explain quantum computing in simple terms"
+ )
>> Starting multi-LLM execution on ALL available models via io.net API
** Prompt: Explain quantum computing in simple terms
** Fetching current model list from io.net API...
** Found 17 models from API
** Successfully fetched 17 models from io.net API
** Randomly selecting 1 models from 17 available models
** Random selection completed
** Selected models ( 1 ):
   1 .  mistralai/Magistral-Small-2506
>> Executing 1 models sequentially
  Processing model 1 / 1 : mistralai/Magistral-Small-2506
>> Multi-LLM execution completed in 2.71 seconds
** Success rate: 100 %
> print(result_one$summary)
$total_models
[1] 1

$successful_models
[1] 1

$failed_models
[1] 0

$success_rate
[1] 100

$total_execution_time
[1] 2.71214

$avg_model_time
[1] 2.710538

$min_model_time
[1] 2.710538

$max_model_time
[1] 2.710538

$total_tokens_used
[1] 306

$failed_details
list()

$failed_model_names
character(0)

> print_multiLLM_results(result_one)
[OK] mistralai/Magistral-Small-2506:
Sure! Here's a simple explanation of **quantum computing**:

### What is Quantum Computing?
Quantum computing is a type of computing that uses **quantum physics** (the science of tiny particles like atoms and electrons) to solve problems much faster than regular computers in certain cases.

### Key Differences from Regular Computers:
1. **Bits vs. Qubits**:
   - Regular computers use **bits** (0 or 1) as their smallest unit of information.
   - Quantum computers use **qubits**, which can be **0, 1, or both at the same time** (thanks to a property called **superposition**).

2. **Parallel Processing**:
   - A regular computer checks possibilities one at a time.
   - A quantum computer checks **many possibilities at once** (using **entanglement**), making it super fast for complex problems like:
     - Breaking encryption
     - Simulating molecules for drug discovery
     - Optimizing traffic or financial systems

### Why Is It Important?
- It can solve problems that would take a regular computer **thousands of years** in seconds or minutes.
- Still in early stages, but promising for science, medicine, and cybersecurity.

### Analogy:
Imagine you're in a maze. A regular computer checks each path one by one. A quantum computer explores **all paths at once** and finds the exit instantly.

Would you like a fun fact or example? 😊

> # Run first 5 models in list order (no random selection)
> result_top5 <- multiLLMviaionet(
+   prompt = "What is machine learning?",
+   max_models = 5,
+   random_selection = FALSE
+ )
>> Starting multi-LLM execution on ALL available models via io.net API
** Prompt: What is machine learning?
** Using cached model list ( 17  models, cached  0  minutes ago)
>>  Limiting to first 5 models from specified list
** Selected models ( 5 ):
   1 .  moonshotai/Kimi-K2-Thinking
   2 .  zai-org/GLM-4.6
   3 .  moonshotai/Kimi-K2-Instruct-0905
   4 .  meta-llama/Llama-3.2-90B-Vision-Instruct
   5 .  openai/gpt-oss-120b
>> Executing 5 models in parallel (async)
Loading required package: deepRstudio
Loading required package: chatAI4R
Loading required package: deepRstudio
Loading required package: deepRstudio
Loading required package: deepRstudio
Loading required package: deepRstudio
Loading required package: deepRstudio
Loading required package: chatAI4R
Loading required package: chatAI4R
Loading required package: chatAI4R
Loading required package: chatAI4R
Loading required package: chatAI4R
>> Monitoring async execution progress:
   [OK] moonshotai/Kimi-K2-Instruct-0905 completed ( 1 / 5 )
   [OK] moonshotai/Kimi-K2-Thinking completed ( 2 / 5 )
   [OK] openai/gpt-oss-120b completed ( 3 / 5 )
   [OK] zai-org/GLM-4.6 completed ( 4 / 5 )
[WARN] Monitor timed out after 120 seconds. Marking remaining futures as timed out.
>> All 5 models completed asynchronously
>> Failed models (reason):
    [FAIL] meta-llama/Llama-3.2-90B-Vision-Instruct - Future execution failed: $ operator is invalid for atomic vectors
>> Multi-LLM execution completed in 123.13 seconds
** Success rate: 80 %
> print_multiLLM_results(result_top5)
[OK] moonshotai/Kimi-K2-Instruct-0905:
Machine learning (ML) is a sub-field of artificial intelligence that gives computers the ability to learn patterns from data without being explicitly programmed with rigid, hand-written rules.

Key ideas in one minute
1. Data instead of rules: Feed the algorithm many examples (images, text, numbers, sensor readings, …).
2. Pattern extraction: The algorithm automatically finds the statistical regularities that map inputs to outputs (e.g., pixel patterns → “cat”; credit-history patterns → “default risk 0.07”).
3. Generalization: After training, the model can predict on new, unseen data that follow the same distribution.
4. Feedback loop: Predictions can be evaluated and the model updated continuously, so the system keeps improving.

Formal definition (Tom Mitchell, 1997)
“A computer program is said to learn from experience E with respect to some task T and performance measure P, if its performance on T, as measured by P, improves with experience E.”

Typical workflow
1. Collect & label data → 2. Choose model family (linear, tree, neural net, …) → 3. Train (optimize parameters to minimize error) → 4. Validate & tune → 5. Deploy & monitor.

Major families
• Supervised: labeled examples (spam/not-spam, house prices).
• Unsupervised: no labels, find structure (clusters, anomalies, embeddings).
• Reinforcement: learn by trial-and-error rewards (game playing, robotics).
• Semi/self-supervised: mix of labeled and unlabeled data.

Why it matters
Turns raw data into actionable predictions or decisions at scale: recommendation engines, fraud detection, machine translation, medical imaging, autonomous vehicles, scientific discovery, etc.

[OK] moonshotai/Kimi-K2-Thinking:
 Machine learning is a branch of artificial intelligence that enables computers to learn patterns from data and make decisions or predictions without being explicitly programmed for every specific task.

**Core Concept:**
Instead of writing detailed step-by-step instructions, you provide the computer with examples (data) and let it discover the underlying rules itself. It's like teaching a child to recognize a cat: rather than listing every possible feature, you show them many pictures of cats until they learn to identify one on their own.

**Main Types:**

*   **Supervised Learning:** The algorithm learns from labeled examples (e.g., emails marked as "spam" or "not spam") to predict labels for new data.
*   **Unsupervised Learning:** The algorithm finds hidden patterns in unlabeled data (e.g., grouping customers into segments based on purchasing behavior).
*   **Reinforcement Learning:** The algorithm learns through trial and error by receiving rewards or penalties (e.g., a game-playing AI that improves by winning or losing).

**Everyday Examples:**
- Spam email filters
- Netflix recommendations
- Face recognition in photos
- Voice assistants like Siri or Alexa
- Fraud detection in banking

In essence, machine learning turns data into actionable insights, automating complex tasks that would be impossible to program manually.

[OK] openai/gpt-oss-120b:
The user asks "What is machine learning?" Straightforward. Provide definition, overview, types, examples. No disallowed content. Provide concise answer.

[OK] zai-org/GLM-4.6:

<think>1.  **Deconstruct the User's Request:**
    *   The user is asking a fundamental, broad question: "What is machine learning?"
    *   This isn't a simple, one-sentence answer. It requires a layered explanation, starting simple and getting progressively more detailed.
    *   The target audience is likely a beginner, someone who has heard the term but doesn't understand it. I should avoid overly technical jargon at the beginning and introduce it gradually with clear explanations.
    *   I need to structure the answer logically to guide the user from a high-level concept to the specifics.

2.  **Structure the Answer:** I'll plan the sections first.
    *   **The Core Analogy (The Hook):** Start with something relatable. The "learning from experience" analogy is perfect. It's the core idea and immediately grounds the concept. Comparing it to a child learning is very effective.
    *   **The Formal Definition:** After the analogy, give a more formal, but still simple, definition. I'll use Arthur Samuel's classic definition ("the field of study that gives computers the ability to learn without being explicitly programmed") because it's concise and powerful. Then I'll paraphrase it in modern terms.
    *   **How Does It Actually Work? (The "Magic" Explained):** This is the crucial part. I need to break down the process into simple steps.
        *   Step 1: Data (The "Experience"). Emphasize that data is the fuel.
        *   Step 2: The Algorithm (The "Learning Process"). This is the recipe or the method.
        *   Step 3: The Model (The "Result"). This is the "trained brain."
        *   Step 4: Prediction/Inference (The "Using the Knowledge"). This is the output.
    *   **A Simple, Concrete Example:** The "Spam Email Filter" is a classic and universally understood example. I'll walk through the four steps using this example to make it tangible.
        *   Data: Thousands of emails, labeled "spam" or "not spam."
        *   Algorithm: A classification algorithm (like Naive Bayes, though I don't need to name it unless I want to add a "further reading" section).
        *   Model: The trained filter that knows which words/phrases are associated with spam.
        *   Prediction: A new email arrives, and the model classifies it.
    *   **Key Types of Machine Learning:** This is the next level of detail. I need to explain the main categories. For each, I'll provide:
        *   A simple definition.
        *   The key difference (e.g., labeled vs. unlabeled data).
        *   A clear example.
        *   The main categories are:
            *   **Supervised Learning:** The most common and easiest to understand. The "teacher with answers" analogy works well. Examples: spam filter, image classification, price prediction. I'll briefly mention its sub-types (Classification and Regression).
            *   **Unsupervised Learning:** The "no teacher" analogy. The goal is to find patterns. Examples: customer segmentation, anomaly detection. I'll briefly mention its sub-types (Clustering and Association).
            *   **Reinforcement Learning:** The "learning by trial and error" analogy. This is a bit more complex, so I'll use a simple example like a video game AI or a self-driving car learning to park. Key concepts: agent, environment, actions, rewards.
    *   **Why is Machine Learning So Important Now?** This provides context and answers the "why should I care?" question.
        *   Big Data: More fuel (data) than ever before.
        *   Computational Power: More powerful engines (GPUs) to process the data.
        *   Algorithmic Advances: Better recipes (algorithms) have been developed.
    *   **Real-World Applications (Where You See It Every Day):** This section makes it all real. I'll list a diverse set of examples to show the breadth of ML's impact.
        *   Recommendation Engines (Netflix, Amazon)
        *   Virtual Assistants (Siri, Alexa)
        *   Self-Driving Cars
        *   Medical Diagnosis
        *   Fraud Detection
        *   Language Translation
    *   **Summary/Conclusion:** A final, concise paragraph to tie everything together. Reiterate the core idea: learning from data to make predictions or decisions.

3.  **Refine the Language:**
    *   Use analogies and metaphors (learning from experience, teacher, recipe, brain, fuel). These are powerful teaching tools.
    *   Use bolding to highlight key terms (`data`, `algorithm`, `model`, `Supervised Learning`, etc.). This helps with scannability and retention.
    *   Use bullet points and numbered lists to break up large blocks of text and make

[FAIL] meta-llama/Llama-3.2-90B-Vision-Instruct:
[no response] - Future execution failed: $ operator is invalid for atomic vectors

> # Random selection of 10 models from all available
> result_random10 <- multiLLMviaionet(
+   prompt = "Write a Python function to calculate fibonacci numbers",
+   max_models = 10,
+   random_selection = TRUE,
+   temperature = 0.3,
+   streaming = FALSE
+ )
>> Starting multi-LLM execution on ALL available models via io.net API
** Prompt: Write a Python function to calculate fibonacci numbers
** Using cached model list ( 17  models, cached  4  minutes ago)
** Randomly selecting 10 models from 17 available models
** Random selection completed
** Selected models ( 10 ):
   1 .  mistralai/Magistral-Small-2506
   2 .  Qwen/Qwen2.5-VL-32B-Instruct
   3 .  meta-llama/Llama-3.2-90B-Vision-Instruct
   4 .  Intel/Qwen3-Coder-480B-A35B-Instruct-int4-mixed-ar
   5 .  meta-llama/Llama-4-Maverick-17B-128E-Instruct-FP8
   6 .  openai/gpt-oss-120b
   7 .  moonshotai/Kimi-K2-Instruct-0905
   8 .  Qwen/Qwen3-Next-80B-A3B-Instruct
   9 .  moonshotai/Kimi-K2-Thinking
   10 .  deepseek-ai/DeepSeek-R1-0528
>> Executing 10 models in parallel (async)
Loading required package: deepRstudio
Loading required package: deepRstudio
Loading required package: deepRstudio
Loading required package: deepRstudio
Loading required package: deepRstudio
Loading required package: deepRstudio
Loading required package: chatAI4R
Loading required package: chatAI4R
Loading required package: chatAI4R
Loading required package: chatAI4R
Loading required package: chatAI4R
Loading required package: chatAI4R
>> Monitoring async execution progress:
   [OK] mistralai/Magistral-Small-2506 completed ( 1 / 10 )
   [OK] meta-llama/Llama-4-Maverick-17B-128E-Instruct-FP8 completed ( 2 / 10 )
   [OK] moonshotai/Kimi-K2-Instruct-0905 completed ( 3 / 10 )
   [OK] Qwen/Qwen3-Next-80B-A3B-Instruct completed ( 4 / 10 )
   [OK] Intel/Qwen3-Coder-480B-A35B-Instruct-int4-mixed-ar completed ( 5 / 10 )
   [OK] deepseek-ai/DeepSeek-R1-0528 completed ( 6 / 10 )
   [OK] Qwen/Qwen2.5-VL-32B-Instruct completed ( 7 / 10 )
   [OK] meta-llama/Llama-3.2-90B-Vision-Instruct completed ( 8 / 10 )
   [OK] openai/gpt-oss-120b completed ( 9 / 10 )
   [OK] moonshotai/Kimi-K2-Thinking completed ( 10 / 10 )
>> All 10 models completed asynchronously
>> Multi-LLM execution completed in 19.87 seconds
** Success rate: 100 %
> print_multiLLM_results(result_random10)
[OK] mistralai/Magistral-Small-2506:
# Python Function to Calculate Fibonacci Numbers

Here's a Python function that calculates Fibonacci numbers using an efficient iterative approach:

```python
def fibonacci(n):
    """
    Calculate the nth Fibonacci number.

    Args:
        n (int): The position in the Fibonacci sequence (0-based index)

    Returns:
        int: The nth Fibonacci number
    """
    if n < 0:
        raise ValueError("Input must be a non-negative integer")
    elif n == 0:
        return 0
    elif n == 1:
        return 1

    a, b = 0, 1
    for _ in range(2, n + 1):
        a, b = b, a + b
    return b
```

### Alternative Implementations:

1. **Recursive approach (less efficient for large n):**
```python
def fibonacci_recursive(n):
    if n < 0:
        raise ValueError("Input must be a non-negative integer")
    elif n <= 1:
        return n
    return fibonacci_recursive(n-1) + fibonacci_recursive(n-2)
```

2. **Using memoization (optimized recursive approach):**
```python
from functools import lru_cache

@lru_cache(maxsize=None)
def fibonacci_memoized(n):
    if n < 0:
        raise ValueError("Input must be a non-negative integer")
    elif n <= 1:
        return n
    return fibonacci_memoized(n-1) + fibonacci_memoized(n-2)
```

3. **Generator function to yield Fibonacci sequence:**
```python
def fibonacci_sequence(max_n=None):
    a, b = 0, 1
    count = 0
    while max_n is None or count < max_n:
        yield a
        a, b = b, a + b
        count += 1
```

The iterative approach (first implementation) is generally the most efficient for single calculations, with O(n) time complexity and O(1) space complexity. The recursive approach has exponential time complexity, while the memoized version improves this to O(n) time with O(n) space.

[OK] meta-llama/Llama-4-Maverick-17B-128E-Instruct-FP8:
**Fibonacci Function in Python**
=====================================

### Overview

The Fibonacci sequence is a series of numbers where a number is the sum of the two preceding ones, usually starting with 0 and 1.

### Code

```python
def fibonacci(n: int) -> int:
    """
    Calculate the nth Fibonacci number.

    Args:
    n (int): The position of the Fibonacci number to calculate.

    Returns:
    int: The nth Fibonacci number.

    Raises:
    ValueError: If n is a negative integer.
    """

    if n < 0:
        raise ValueError("n must be a non-negative integer")
    elif n <= 1:
        return n
    else:
        return fibonacci(n-1) + fibonacci(n-2)

def fibonacci_memo(n: int, memo: dict = {}) -> int:
    """
    Calculate the nth Fibonacci number using memoization.

    Args:
    n (int): The position of the Fibonacci number to calculate.
    memo (dict): A dictionary to store previously calculated Fibonacci numbers.

    Returns:
    int: The nth Fibonacci number.

    Raises:
    ValueError: If n is a negative integer.
    """

    if n < 0:
        raise ValueError("n must be a non-negative integer")
    elif n <= 1:
        return n
    elif n not in memo:
        memo[n] = fibonacci_memo(n-1, memo) + fibonacci_memo(n-2, memo)
    return memo[n]

def fibonacci_iterative(n: int) -> int:
    """
    Calculate the nth Fibonacci number iteratively.

    Args:
    n (int): The position of the Fibonacci number to calculate.

    Returns:
    int: The nth Fibonacci number.

    Raises:
    ValueError: If n is a negative integer.
    """

    if n < 0:
        raise ValueError("n must be a non-negative integer")
    elif n <= 1:
        return n

    a, b = 0, 1
    for _ in range(2, n+1):
        a, b = b, a + b

    return b

# Example usage:
if __name__ == "__main__":
    n = 10
    print(f"Fibonacci({n}) = {fibonacci(n)}")
    print(f"Fibonacci({n}) with memoization = {fibonacci_memo(n)}")
    print(f"Fibonacci({n}) iterative = {fibonacci_iterative(n)}")
```

### Explanation

The code provides three different implementations to calculate Fibonacci numbers:

1.  **Recursive Implementation (`fibonacci` function):** This is a straightforward recursive implementation. However, it has an exponential time complexity due to repeated calculations.
2.  **Memoized Implementation (`fibonacci_memo` function):** This implementation uses memoization to store previously calculated Fibonacci numbers, reducing the time complexity to linear.
3.  **Iterative Implementation (`fibonacci_iterative` function):** This implementation uses a loop to calculate Fibonacci numbers iteratively, also achieving linear time complexity.

The example usage demonstrates how to use these functions to calculate the 10th Fibonacci number.

[OK] moonshotai/Kimi-K2-Instruct-0905:
```python
def fibonacci(n: int) -> int:
    """
    Return the n-th Fibonacci number (0-based indexing).

    Parameters
    ----------
    n : int
        Index (non-negative) of the desired Fibonacci number.

    Returns
    -------
    int
        The n-th Fibonacci number.

    Raises
    ------
[…]
