### 2025-12-02 #9283 @tzuV
@kumeS in one of the coming days I will cover the responses you have made to my issues. So I think my review is close to being completed. Thank you for addressing my points!

I see you have covered all my issues except the `discussion_flow functions_v*`. Are you aware of that, or haven't you had time to look into it?
