### 2025-12-12 kumeS/chatAI4R#38 @kumeS
Hi @tzuV,

I can see these functions using `?`.
this is package ‘chatAI4R’ version ‘1.2.6’.

```r
?gemini4R
?geminiGrounding4R
?interpretResult
?multiLLMviaionet
```
