### 2025-12-12 kumeS/chatAI4R#38 @tzuV
@kumeS okay. I'll check it later. I tried to write 'chatAI4R::' but nothing appeared.
