### 2025-12-15 10aee30
v1.1.x: JOSS review
