+++ R/discussion_flow_v1.R
@@ -19,13 +19,11 @@
 #' @description Simulates interactions and flow control between three different roles of LLM-based bots (LLBs).
 #' @param issue The issue to be discussed. Example: "I want to solve linear programming and create a timetable."
 #' @param Domain The domain of the discussion, default is "bioinformatics".
-#' @param Model The model to be used, default is "gpt-4o-mini".
+#' @param Model The model to be used, default is "gpt-5-nano".
 #' @param api_key The API key for OpenAI, default is retrieved from the system environment variable "OPENAI_API_KEY".
 #' @param language The language for the discussion, default is "English".
 #' @param Summary_nch The number of characters for the summary, default is 50.
 #' @param verbose Logical, whether to print verbose output, default is TRUE.
-#' @param Nonfuture Logical, whether to use an asynchronous processing or not, default is not to use (TRUE).
-#' @param sayENorJA Logical, whether to say in English or Japanese, default is TRUE. This feature is available on  macOS system only.
 #' @importFrom future plan future multisession resolved
 #' @importFrom igraph graph add_vertices layout_nicely add_edges layout_with_fr
 #' @importFrom deepRstudio is_mac deepel
@@ -34,35 +32,78 @@
 #' @author Satoshi Kume
 #' @examples
 #' \dontrun{
-#' issue <-  "I want to solve linear programming and create a timetable."
+#' issue <-  "I want to discuss about a potentiality of multivariate analysis for metabolomics data."
 #'
 #' #Run Discussion with the domain of bioinformatics
 #' discussion_flow_v1(issue)
 #' }
 
-#issue = "I want to solve linear programming and create a timetable.";Domain = "bioinformatics";Model = "gpt-4o-mini";api_key = Sys.getenv("OPENAI_API_KEY");language = "English";Summary_nch = 50; verbose = TRUE; sayENorJA = FALSE; Nonfuture = TRUE
-
-#discussion_flow_v1(issue, sayENorJA = FALSE)
+#issue = "I want to discuss about a potentiality of multivariate analysis for metabolomics data.";Domain = "bioinformatics";Model = "gpt-5-nano";api_key = Sys.getenv("OPENAI_API_KEY");language = "English";Summary_nch = 50; verbose = TRUE
+#discussion_flow_v1(issue)
 
 discussion_flow_v1 <- function(issue,
                                Domain = "bioinformatics",
-                               Model = "gpt-4o-mini",
+                               Model = "gpt-5-nano",
                                api_key = Sys.getenv("OPENAI_API_KEY"),
                                language = "English",
                                Summary_nch = 50,
-                               verbose = TRUE,
-                               Nonfuture = TRUE,
-                               sayENorJA = TRUE){
+                               verbose = TRUE){
 
 #Create multi-session
+Nonfuture <- TRUE
+sayENorJA = TRUE
 future::plan(future::multisession())
[…]
