### 2025-12-15 456b1c3
v1.1.x: JOSS review
