+++ R/discussion_flow_v2.R
@@ -19,7 +19,7 @@
 #'
 #' @title discussion_flow_v2: Interactions and Flow Control Between LLM-based Bots (LLBs)
 #' @description Simulates interactions and flow control between three different roles of LLM-based bots (LLBs).
-#' @param issue The issue to be discussed. Example: "I want to solve linear programming and create a timetable."
+#' @param issue The issue to be discussed. Example: "I want to perform differential gene expression analysis from RNA-seq data and interpret enriched pathways."
 #' @param Domain The domain of the discussion, default is "bioinformatics".
 #' @param Model The LLM model to be used, default is "gpt-4o-mini".
 #' @param api_key The API key for OpenAI, default is retrieved from the system environment variable "OPENAI_API_KEY".
@@ -29,7 +29,6 @@
 #' @param R_expert_setting Logical, whether R expert settings are enabled, default is TRUE.
 #' @param verbose Logical, whether to print verbose output, default is TRUE.
 #' @param sayENorJA Logical, whether to speak in English or Japanese, default is TRUE. This feature is available on macOS systems only.
-#' @param rep_x Numeric, a number of repeat for the conversations.
 #' @param rep_x Numeric, a number of repeats for the conversations, default is 3.
 #' @importFrom future plan future multisession resolved
 #' @importFrom igraph graph add_vertices layout_nicely add_edges layout_with_fr
@@ -39,13 +38,13 @@
 #' @author Satoshi Kume
 #' @examples
 #' \dontrun{
-#' issue <-  "I want to solve linear programming and create a timetable."
+#' issue <-  "I want to perform differential gene expression analysis from RNA-seq data and interpret enriched pathways."
 #'
 #' #Run Discussion with the domain of bioinformatics
 #' discussion_flow_v2(issue)
 #' }
 
-#issue = "I want to solve linear programming and create a timetable.";Domain = "bioinformatics";Model = "gpt-4o-mini";api_key = Sys.getenv("OPENAI_API_KEY");language = "English";Summary_nch = 50; verbose = TRUE; sayENorJA = FALSE; Sentence_difficulty = 2; R_expert_setting = TRUE; rep_x = 3
+#issue = "I want to perform differential gene expression analysis from RNA-seq data and interpret enriched pathways.";Domain = "bioinformatics";Model = "gpt-4o-mini";api_key = Sys.getenv("OPENAI_API_KEY");language = "English";Summary_nch = 50; verbose = TRUE; sayENorJA = FALSE; Sentence_difficulty = 2; R_expert_setting = TRUE; rep_x = 3
 #discussion_flow_v2(issue, sayENorJA = FALSE)
 
 discussion_flow_v2 <- function(issue,
@@ -64,11 +63,55 @@ discussion_flow_v2 <- function(issue,
 future::plan(future::multisession())
 DEEPL <- any(names(Sys.getenv()) == "DeepL_API_KEY")
 
+normalize_text <- function(x) {
+  if (is.null(x)) {
+    return("")
+  }
+
+  if (is.data.frame(x)) {
+    if ("content" %in% names(x)) {
[…]
