### 2025-12-24 68a70de
v1.1.x: JOSS review
