### 2025-12-24 kumeS/chatAI4R#27 @kumeS
## https://github.com/kumeS/chatAI4R/commit/68a70de076368c2d1474630585728111044dc067

I apologize for the delay in handling this case.
In this response, I have removed Japanese videos URLs from the README.md.
After the JOSS review is complete, I will create an English version of the video and upload it to Youtube.
Best regards,
