### 2025-12-24 1d927d5
v1.1.x: JOSS review
