### 2025-12-24 kumeS/chatAI4R#40 @kumeS
Hi @khoa-yelo ,

## Updates: https://github.com/kumeS/chatAI4R/commit/1d927d5b5a790bfc48dc85352477a71752ef9733

I've updated the manuscript to include integration with Plumber and maintenance of multiple APIs. This revision clarifies the target audience and the purpose of the package.
