### 2025-12-24 kumeS/chatAI4R#24 @kumeS
Hi @khoa-yelo,
I will do a final check of the package and submit it to CRAN tomorrow.
Best regards, Satoshi
