### 2025-12-25 3558f18
v1.1.x: JOSS review
