+++ man/discussion_flow_v2.Rd
@@ -69,7 +69,8 @@ English is recommended as the input language, but the review will also be conduc
 }
 \examples{
 \dontrun{
-issue <-  "I want to perform differential gene expression analysis from RNA-seq data and interpret enriched pathways."
+issue <-  "I want to perform differential gene expression analysis from
+           RNA-seq data and interpret enriched pathways."
 
 #Run Discussion with the domain of bioinformatics
 discussion_flow_v2(issue)
