+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: chatAI4R
 Type: Package
 Title: Chat-Based Interactive Artificial Intelligence for R
-Version: 1.2.9
+Version: 1.3.0
 Date: 2025-11-30
 Authors@R: c(
     person (given = "Satoshi", family = "Kume",
@@ -13,8 +13,7 @@ Maintainer: Satoshi Kume <satoshi.kume.1984@gmail.com>
 Description: The Large Language Model (LLM) represents a groundbreaking advancement 
     in data science and programming, and also allows us to extend the world of R. 
     A seamless interface for integrating the 'OpenAI' Web APIs into R is provided in this package. 
-    This package leverages LLM-based AI techniques, enabling efficient knowledge discovery and data analysis 
-    (see 'OpenAI' Web APIs details <https://openai.com/blog/openai-api>).
+    This package leverages LLM-based AI techniques, enabling efficient knowledge discovery and data analysis.
     The previous functions such as seamless translation and image generation have been moved 
     to other packages 'deepRstudio' and 'stableDiffusion4R'.
 Depends: R (>= 4.2.0)
