### 2025-12-25 5ae17fb
v1.x: JOSS review
