### 2025-12-27 #9283 @kumeS
@crvernon @khoa-yelo
Regarding https://github.com/kumeS/chatAI4R/issues/24,  the CRAN team is on their winter break, and will be closed until Jan 7, 2026. I will submit the package after the new year break.
Regarding https://github.com/kumeS/chatAI4R/issues/40, I have updated the article to clarify the audience and its application as package users.
