### 2026-01-07 661b35a
v1.x: JOSS review
