+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: chatAI4R
 Type: Package
 Title: Chat-Based Interactive Artificial Intelligence for R
 Version: 1.3.1
-Date: 2025-11-30
+Date: 2026-1-8
 Authors@R: c(
     person (given = "Satoshi", family = "Kume",
     role = c("aut", "cre"), 
