### 2026-01-12 a6a01d4
v1.x: JOSS review
