+++ README.md
@@ -38,10 +38,10 @@ chatAI4R is an experimental project aimed at developing and implementing various
 
 ### 🏗️ 4-Layer Architecture (Beyond Simple API Wrapper)
 
-**Layer 1 - Core**: Direct API access (`chat4R`, `gemini4R`, `multiLLMviaionet`)
-**Layer 2 - Usage**: Conversation memory, text processing (`conversation4R`, `proofreadText`)
-**Layer 3 - Workflow**: Multi-bot systems, R package automation (`discussion_flow_v1`, `autocreateFunction4R`)
-**Layer 4 - Expertise**: Domain-specific analysis (`interpretResult` with 13 domains, `geminiGrounding4R`)
+**Layer 1 - Core Functions**: Direct API access (`chat4R`, `gemini4R`, `multiLLMviaionet`)
+**Layer 2 - Advanced Functions**: Intelligent processing and domain-specific analysis (`interpretResult`, `conversation4R`, `geminiGrounding4R`)
+**Layer 3 - Workflow Functions**: Multi-agent collaboration and package automation (`discussion_flow_v2`, `autocreateFunction4R`)
+**Layer 4 - Integration Functions**: Ecosystem connectivity (deploying R-based web APIs via `plumber` for backend embedding, and GUI support)
 
 ### 🔒 Production-Ready Security
 - HTTP status validation and error handling
@@ -182,8 +182,7 @@ conversation4R("It has 10,000 samples and 50 features")
 
 # 3. R package development
 autocreateFunction4R(
-  FunctionName = "calculate_fold_change",
-  specification = "Calculate log2 fold change with error handling"
+  Func_description = "Create a function to calculate log2 fold change with error handling"
 )
 # → Generates function + Roxygen2 documentation
 ```
@@ -191,17 +190,18 @@ autocreateFunction4R(
 ### Advanced Features
 
 **Statistical Interpretation** (13+ domains):
+
 ```r
 model <- lm(mpg ~ wt + hp, data = mtcars)
-interpretResult(summary(model), domain = "Regression Analysis")
+interpretResult(analysis_type = "regression", result_text = summary(model))
 ```
 
 **Multi-Expert Discussion**:
+
 ```r
 discussion_flow_v1(
-  title = "RNA-seq analysis strategy",
-  expert1 = "Bioinformatics Specialist",
-  expert2 = "Statistical Geneticist"
+  issue = "RNA-seq analysis strategy",
+  Domain = "Bioinformatics"
 )
 ```
 
@@ -227,27 +227,6 @@ chatAI4R includes a comprehensive automated testing framework:
 - **Base Tests**: 14 core functions for quick validation (~9 seconds)
 - **Extended Tests**: Phase 1-5 suites covering all function categories (~50 seconds)
 - **CI/CD Ready**: Automated testing on Windows, macOS, and Linux via GitHub Actions
-
-### Running Tests
-
-```r
-# Quick validation (recommended for daily development)
-source("tests/test/run_all_tests.R")
-results <- run_all_tests()
-
-# Comprehensive testing (before releases)
-source("tests/test/phase1_tests.R")
-p1_results <- run_phase1_tests()
-# ... continue with phase2-5
-```
-
-See [tests/test/README.md](https://github.com/kumeS/chatAI4R/blob/main/tests/test/README.md) for detailed documentation.
-
-- [examples for proodread English text and check error in Japanese on Rstudio](https://youtu.be/VJaltAS9Ef8)
-
-###  Applied usage of the chatAI4R package
-
-- AI-based chatting loaded with highly-technical documents (RIKEN Pressrelease text)
 
 ## Prompts for chatGPT / GPT-4
 
[…]
