+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: chatAI4R
 Type: Package
 Title: Chat-Based Interactive Artificial Intelligence for R
-Version: 1.3.1
+Version: 1.3.2
 Date: 2026-1-8
 Authors@R: c(
     person (given = "Satoshi", family = "Kume",
