### 2026-01-13 dd2c689
v1.x: JOSS review
