+++ README.md
@@ -500,8 +500,8 @@ Kume S. (2025) chatAI4R: Chat-based Interactive Artificial Intelligence for R. V
 @misc{Kume2025chatAI4R,
   title={chatAI4R: Chat-Based Interactive Artificial Intelligence for R},
   author={Kume, Satoshi},
-  year={2025},
-  version={1.1.0},
+  year={2026},
+  version={1.3.1},
   publisher={GitHub},
   note={R Package with Multi-LLM Capabilities},
   howpublished={\url{https://github.com/kumeS/chatAI4R}},
