### 2026-01-13 1e79063
v1.x: JOSS review
