+++ paper/paper.md
@@ -56,15 +56,17 @@ The package maintains production-level reliability through CRAN distribution, en
 
 The chatAI4R package implements a four-layered application architectural design that provides progressive functionality. This modular approach ensures accessibility for all users.
 
-The Core Functions (Layer 1) establish unified API interfaces for nine LLM platforms. The `chat4R()`, `gemini4R()`, and `multiLLMviaionet()` functions provide standardized access patterns while handling authentication, rate limiting, and error management.
+The Core Functions (Layer 1) establish unified API interfaces for multiple LLM platforms. The `chat4R()`, `chat4R_history()`, `gemini4R()`, `multiLLMviaionet()`, and related functions provide standardized access patterns while handling authentication, rate limiting, and error management.
 
-The Advanced Functions (Layer 2) introduce R-specific capabilities through intelligent prompt engineering. The `interpretResult()` function employs specialized templates for 13 analysis types, automatically generating domain-appropriate interpretations.
+The Advanced Functions (2nd Layered Functions, Usage/Task) enable execution of simple LLM tasks through intelligent prompt engineering. Functions such as `conversation4R()`, `revisedText()`, `enrichTextContent()`, `TextSummary()`, and `multiLLMviaionet()` provide chat memory management, translation, proofreading, and text processing capabilities.
 
-The Workflow Functions (Layer 3) implement the multi-agent collaboration system. The `discussion_flow_v1()` and `discussion_flow_v2()` architectures employ role-based prompt engineering where each agent maintains distinct personas, enabling iterative solution refinement.
+The Workflow Functions (3rd Layered Functions, Workflow) focus on supporting R package development. The `createSpecifications4R()`, `convertRscript2Function()`, `addRoxygenDescription()`, and `convertScientificLiterature()` functions provide comprehensive development automation, transforming LLMs into powerful assistants tailored for R programming workflows.
 
-The Integration Functions (Layer 4) provide connectivity with R ecosystem tools and development workflows. This layer facilitates the deployment of R-based web APIs by combining `chatAI4R` with the `plumber` package, allowing users to implement LLM-powered backend processing embedded in external platforms. Additionally, it supports GUI-based interactions, ensuring accessibility for users who prefer graphical interfaces over command-line operations.
+The Integration Functions (4th Layered Functions, Expertise) provide data mining and advanced analysis capabilities, while also providing connectivity with R ecosystem tools and development workflows. The `interpretResult()` function employs specialized templates for 13 analysis types, automatically generating domain-appropriate interpretations. The `extractKeywords()` extracts keywords from text, `vision4R()` enables image analysis via Vision API, and the multi-agent discussion system (`discussion_flow_v1()`, `discussion_flow_v2()`) employs role-based prompt engineering where each agent maintains distinct personas, enabling iterative solution refinement.
 
-The package employs defensive programming with comprehensive error handling, input validation, and graceful degradation when API services are unavailable. Package reliability is ensured through automated testing. Furthermore, to address the challenge of frequent API changes across multiple providers (OpenAI, Gemini, io.net, Replicate, Dify), chatAI4R employs a sustainable maintenance strategy combining user contributions with high-frequency, LLM-assisted autonomous code updates, ensuring long-term resilience.
+The package supports integration with six API platforms: OpenAI, Google Gemini, DeepL translation, Replicate, Dify, and IO.net Intelligence API. This comprehensive API connectivity ensures flexibility in model selection and deployment scenarios. The package also facilitates the deployment of R-based web APIs by combining `chatAI4R` with the `plumber` package, allowing users to implement LLM-powered backend processing embedded in external platforms. Additionally, it supports GUI-based interactions, ensuring accessibility for users who prefer graphical interfaces over command-line operations.
+
[…]
