### 2026-01-13 419bbf3
v1.x: JOSS review
