### 2026-01-23 #9283 @crvernon
@kumeS your last reviewer [just signed off last week](https://github.com/openjournals/joss-reviews/issues/9283#issuecomment-3742154617).  I ask for your patience as we manage a very large queue of submissions and deal with those needs as they arise in order.
