### 2026-01-23 #9283 @crvernon
👋 @kumeS - I need you to make the following changes in your paper:

- [x] Your figure 1 caption echos "Figure 1:"
- [x] Your code examples in the "Usage" section flow off of the page.  I would remove the usage section completely and simply mention that examples are provided in your [documentation](https://kumes.github.io/chatAI4R/).
- [x] LINE 128:  the "r" as in the R programming language should be capitalized here.  You can do this in your bib file by placing characters you wish to maintain the formatting of in curly brackets.
- [x] LINE 130:  the "r" as in the R programming language should be capitalized.
- [x] LINE 132:  the "r" as in the R programming language should be capitalized.
- [x] LINE 137:  the "r" as in the R programming language should be capitalized.

- [x]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted in your code repository.

- [x]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly including the case) and the author list (should match the people and the order in which they are listed on your paper exactly).

- [x]  Please respond with the DOI of the archived version here along with the version tag name released.
