### 2026-01-27 #9283 @kumeS
Hi @crvernon, thank you very much for the updates.
I will respond to all of your comments this weekend.
Many thanks,
Satoshi
