+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @misc{Iegor2024,
-  title = {openai: R Wrapper for OpenAI API},
+  title = {openai: {R} Wrapper for OpenAI API},
   author = {Iegor Rudnytskyi},
   year = {2024},
   note = {R package version 0.4.1,
@@ -10,7 +10,7 @@
 }
 
 @misc{Wanjun2024,
-  title = {gptr: A Convenient R Interface with the OpenAI 'ChatGPT' API},
+  title = {gptr: A Convenient {R} Interface with the OpenAI 'ChatGPT' API},
   author = {Wanjun Gu},
   year = {2024},
   url = {https://cran.r-project.org/web/packages/gptr/index.html},
@@ -24,11 +24,11 @@
   publisher = {The Open Journal}, 
   volume = {10}, number = {105}, 
   pages = {7211}, author = {Hause Lin and Tawab Safi}, 
-  title = {ollamar: An R package for running large language models}, journal = {Journal of Open Source Software} 
+  title = {ollamar: An {R} package for running large language models}, journal = {Journal of Open Source Software} 
 }
 
 @misc{gruber2024rollamarpackageusing,
-      title={rollama: An R package for using generative large language models through Ollama}, 
+      title={rollama: An {R} package for using generative large language models through Ollama}, 
       author={Johannes B. Gruber and Maximilian Weber},
       year={2024},
       eprint={2404.07654},
@@ -63,7 +63,7 @@
 }
 
 @misc{Rodriguez2023,
-  title = {chatgpt: Interface to 'ChatGPT' from R},
+  title = {chatgpt: Interface to 'ChatGPT' from {R}},
   author = {Juan Cruz Rodriguez},
   year = {2023},
   note = {R package version 0.2.3},
