+++ paper/paper.bib
@@ -18,13 +18,14 @@
 }
 
 @article{Lin2025,
+  title = {ollamar: An {R} package for running large language models}
   doi = {10.21105/joss.07211}, 
   url = {https://doi.org/10.21105/joss.07211}, 
   year = {2025}, 
   publisher = {The Open Journal}, 
   volume = {10}, number = {105}, 
-  pages = {7211}, author = {Hause Lin and Tawab Safi}, 
-  title = {ollamar: An {R} package for running large language models}, journal = {Journal of Open Source Software} 
+  pages = {7211}, author = {Hause Lin and Tawab Safi},  
+  journal = {Journal of Open Source Software} 
 }
 
 @misc{gruber2024rollamarpackageusing,
@@ -38,7 +39,6 @@
       doi={10.48550/arXiv.2404.07654}
 }
 
-
 @misc{openai2024gpt4technicalreport,
       title={GPT-4 Technical Report}, 
       author={OpenAI OpenAI},
@@ -49,9 +49,6 @@
       url={https://arxiv.org/abs/2303.08774},
       doi={10.48550/arXiv.2303.08774}
 }
-
-
-
 
 @misc{Wickham2025,
   title = {ellmer: Chat with Large Language Models},
