+++ paper/paper.bib
@@ -18,7 +18,7 @@
 }
 
 @article{Lin2025,
-  title = {ollamar: An {R} package for running large language models}
+  title = {ollamar: An {R} package for running large language models},
   doi = {10.21105/joss.07211}, 
   url = {https://doi.org/10.21105/joss.07211}, 
   year = {2025}, 
