+++ review.txt
@@ -1,0 +1,146 @@
+================================================================================
+JOSS Review Analysis - chatAI4R Package
+================================================================================
+日付: 2026年1月31日
+================================================================================
+
+■ レビュアーからの要求事項
+--------------------------------------------------------------------------------
+
+1. タイトルの特殊フォーマットの確認
+2. GitHub リリースの実施とZenodo等へのアーカイブ
+3. アーカイブのメタデータ確認（タイトル・著者リスト）
+4. アーカイブ版のDOIとバージョンタグ名の報告
+
+================================================================================
+■ 解析結果
+================================================================================
+
+【1】タイトルの確認
+--------------------------------------------------------------------------------
+paper.mdのタイトル:
+  'chatAI4R: Interactive Artificial Intelligence toolkit for Data Science in R'
+
+ステータス: ✅ OK
+- バッククォート（`）やその他の特殊フォーマットは使用されていません
+- タイトルはプレーンテキストで記載されています
+
+--------------------------------------------------------------------------------
+【2】著者リストの確認
+--------------------------------------------------------------------------------
+paper.mdの著者情報:
+  - 名前: Satoshi Kume
+  - ORCID: 0000-0001-7481-2843
+  - corresponding: true
+  - 所属: Bio"Pack"athon, Osaka, Japan
+
+DESCRIPTIONファイルの著者情報:
+  - 名前: Satoshi Kume
+  - Role: aut, cre (Author, Creator)
+  - Email: satoshi.kume.1984@gmail.com
+  - ORCID: 0000-0001-7481-2843
+
+ステータス: ✅ 一致
+- 著者名は一致しています
+- ORCIDも一致しています
+
+--------------------------------------------------------------------------------
+【3】ライセンスの確認
+--------------------------------------------------------------------------------
+LICENSEファイル: Artistic License 2.0
+DESCRIPTIONファイル: Artistic-2.0
+
+ステータス: ✅ 一致
+- Zenodoアーカイブ時に「Artistic License 2.0」を選択する必要があります
+
+================================================================================
+■ 必要なアクション（チェックリスト）
+================================================================================
+
+□ 1. GitHubリリースの作成
+   - 現在のレビュー済みバージョンでGitHubリリースを作成
+   - 推奨バージョンタグ: v1.3.2（DESCRIPTIONに記載のバージョンと一致）
+   - リリースノートにJOSS論文の情報を含める
+
+□ 2. Zenodoへのアーカイブ
+   - GitHubリポジトリをZenodoに連携（未実施の場合）
+   - または手動でZenodoにアップロード
+   
+□ 3. Zenodoメタデータの確認・設定
+   【重要】以下を正確に設定すること：
+   
+   Title（タイトル）- 大文字小文字も含めて完全一致が必要:
+   ┌────────────────────────────────────────────────────────────────────┐
+   │ chatAI4R: Interactive Artificial Intelligence toolkit for Data    │
+   │ Science in R                                                      │
+   └────────────────────────────────────────────────────────────────────┘
+   
+   Authors（著者）:
+   ┌────────────────────────────────────────────────────────────────────┐
+   │ Satoshi Kume                                                      │
+   │ ORCID: 0000-0001-7481-2843                                        │
+   │ Affiliation: Bio"Pack"athon, Osaka, Japan                         │
+   └────────────────────────────────────────────────────────────────────┘
+   
+   License（ライセンス）:
+   ┌────────────────────────────────────────────────────────────────────┐
+   │ Artistic License 2.0                                              │
+   └────────────────────────────────────────────────────────────────────┘
+
[…]
