### 2026-01-31 #9283 @kumeS
@crvernon Thank you for the review.

I have completed the following steps:

1. Additional comments addressed:
I have also addressed your comments with the following changes:
   - Revised the Figure legend
   - Removed the Usage section from `paper.md` and migrated its content to the README
   - Changed lowercase "r" to uppercase "R" in references

These changes can be viewed in the following commit: https://github.com/kumeS/chatAI4R/commit/a28ddcc2f97facbc5e5fb6d6892a9bef99e7cc09

2. Title confirmation: The paper title in paper.md contains no special formatting (no backticks or other special characters).

3. GitHub Release: I have created a GitHub release with version tag `v1.3.2`.
https://github.com/kumeS/chatAI4R/releases/tag/v1.3.2

4. Zenodo Archive: The software has been archived on Zenodo with the following metadata:
   - Title: "chatAI4R: Interactive Artificial Intelligence toolkit for Data Science in R" (matches the paper title exactly)
   - Author: Satoshi Kume (matches the paper)
   - License: Artistic License 2.0 (same as the repository)

4. Archive DOI: https://doi.org/10.5281/zenodo.18438864
   Version tag: v1.3.2

Please proceed with accepting the submission.

Best regards,

Satoshi
