### 2026-02-10 #9283 @kumeS
Hi @crvernon

I am very glad to hear that you had a chance to review my latest update to this JOSS article.
Best regards,

Satoshi
