### 2026-03-27 #9283 @kumeS
Hi, @crvernon @arfon

This is a kind reminder to finalize this JOSS article.
I’m very glad you had a chance to review my latest update to the JOSS article.
Best regards,

Satoshi
